==> Rmarkdowns/1_survey_data_wrangling_sr.Rmd <==
---
title: "Survey Data- Wrangling, Cleaning, Filtering"
author: "Sydney Rilum"
date: "12/15/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# Load necessary packages
library(tidyverse)
library(here)
library(janitor)
library(data.table)  # advanced pivot_table tools
library(naniar)      # replace NA tools
library(sf)          # read in shapefiles
#library(tmap)        # interactive maps
library(zipcodeR)    # for zipcode assignment by county
#library(raster)      # for mapping raster layers
#library(dplyr)       # needed ONLY if working on a windows computer
library(ggwordcloud) # create wordcloud plots
library(wordcloud)
library(patchwork)   # for stacking plots
library(kableExtra)  # for making tables

#library(conflicted)  # allows for choosing preference over conflicting functions from different packages
#conflict_prefer("select", "dplyr") # use select function from tidyverse/dplyr package > from raster package
#conflict_prefer("filter", "dplyr")
#conflict_prefer("summarize", "dplyr")
#conflict_prefer("group_by", "dplyr")
```

# Data Set-up & Cleaning 

```{r}
# Read in combined survey data
survey_data <- read.csv(here("data","survey_data","combined_survey_responses.csv"))

survey_data_tidy <- survey_data %>%
  clean_names() %>% 
  slice(-1) # remove first row of df
```

```{r}
# Convert dataframe to tidydata format (Reshaping the data)

# convert data frame into a data table
survey_data_tidy <- as.data.table(survey_data_tidy)

# use melt() from `datatable` package (advanced version of pivot_longer()) to reshape data from wide to long format
# unlike pivot_longer(), melt() allows for pivoting data tables with both numeric and character data!
survey_data_tidy_long <- survey_data_tidy %>% 
  melt(id.vars = c("response_id", "response_type", "finished"), 
       variable.name = "question", 
       value.name = "answer")
```

## Filter for Santa Barbara & Ventura County Zipcodes 

```{r}
# Q37 - Zipcode Data

# Create a dataframe containing only zipcodes from Ventura and Santa Barbara Counties

## Not run:
download_zip_data() # uses zipcodeR data 
## End(Not run)

ventura_co_zip_full <-
  search_county("ventura", "CA") 

ventura_zipcodes<- 
  ventura_co_zip_full[,c("zipcode", "county", "lat", "lng")]

sb_co_zip_full <- 
 search_county("Santa Barbara", "CA")

sb_zipcodes<- 
  sb_co_zip_full[,c("zipcode", "county", "lat", "lng")]

# combine SB & Ventura zipcode data frames
sb_ventura_co_zip <- 
  rbind(ventura_zipcodes, sb_zipcodes)
```

```{r}
# select survey responses from Ventura and Santa Barbara Counties ONLY
survey_data_tidy_sbv <- merge(survey_data_tidy, sb_ventura_co_zip, 
                              by.x = "q37", by.y = "zipcode")

### Convert dataframe to tidydata format (Reshaping the data)

# convert data frame into a data table
survey_data_tidy_sbv <- as.data.table(survey_data_tidy_sbv)

# use melt() from `datatable` package (advanced version of pivot_longer()) to reshape data from wide to long format
survey_data_tidy_sbv_long <- survey_data_tidy_sbv %>% 
  melt(id.vars = c("response_id", "response_type", "finished", "duration_in_seconds", "county", 
                   "q31", "q32", "q33", "q34", "q35", "q36", "q37", "q38", "q39", "q39_7_text", "q40", "q40_8_text"), 
       variable.name = "question", 
       value.name = "answer")
```

## Filter Survey Response Type

Response Types Key:

0 = speeders (time removal)
1 = completes (approved by our team)
2 = removed by our team for replacement (flagged for not answering ranking questions, but otherwise fine)
3 = response collected via our own outreach (QRcode/email/social media)

(Note: Qualtrics Paid Responses = 0,1,2)

```{r}
# count number of:

# 0 (speeders)
response_0 <- survey_data_tidy %>% 
  mutate(duration_in_seconds = as.numeric(duration_in_seconds)) %>% 
  filter(duration_in_seconds < 180)         # 33 speeders
  
# 1 (completes)
response_1 <- survey_data_tidy %>% 
  filter(response_type == 1)         # 302 completes

# 2 (replaced)
response_2 <- survey_data_tidy %>% 
  filter(response_type == 2)         # 48 extra completes

# 3 (responses collected via email/social media)
response_3 <- survey_data_tidy %>% 
  filter(response_type == 3)         # 146 responses via self-distribution

# 3 completes (finished survey)
response_3_completes <- survey_data_tidy %>% 
  filter(response_type == 3) %>% 
  filter(finished == TRUE)           # 110 completes

# 3 incompletes (did not finish survey)  
response_3_incompletes <- survey_data_tidy %>% 
  filter(response_type == 3) %>% 
  filter(finished == FALSE)          # 36 incompletes

# Total number of responses = 511
# Total number of responses (-speeders) = 496
# Total number of responses (-speeders, -incompletes) = 460

# Total number of SB/Ventura County responses = 400 
# Total number of SBV responses (-speeders) = 364
response_sbv <- survey_data_tidy_sbv %>% 
  mutate(duration_in_seconds = as.numeric(duration_in_seconds)) %>% 
  filter(duration_in_seconds > 180) 

```

```{r}
# remove speeders responses (response type: 0) from survey dataframes
survey_data_tidy_filtered <- survey_data_tidy %>% 
  filter(response_type != 0)

survey_data_tidy_long_filtered <- survey_data_tidy_long %>% 
  filter(response_type != 0)

# remove speeders from SBV datasets
## THESE DATASETS SHOULD BE SAVED AS CSVs (see next code chunk)
survey_data_tidy_sbv_filtered <- survey_data_tidy_sbv %>% 
  filter(duration_in_seconds >= 180) 

survey_data_tidy_sbv_long_filtered <- survey_data_tidy_sbv_long %>% 
  filter(duration_in_seconds >= 180) 
```

## Save Finalized Data Frames as csv 

```{r}
write.csv(survey_data_tidy_sbv_filtered,
          file = here("data", "survey_data_tidy_sbv_filtered.csv"))

write.csv(survey_data_tidy_sbv_long_filtered,
          file = here("data", "survey_data_tidy_sbv_long_filtered.csv"))
```










## Map Survey Response Location by zipcode/county 

```{r}
# Read in California counties outline
#ca_counties <- read_sf(here("data", "ca_counties", "CA_Counties_TIGER2016.shp")) %>% 
#  clean_names()

# Read in US zipcodes tabulation areas
#us_zipcodes <- read_sf(here("data", "zipcode_tabulation_areas_2020", "tl_2021_us_zcta520.shp")) %>% 
#  clean_names()

```

```{r}
#us_zipcodes <- drop_na(us_zipcodes)

# Check class
#class(us_zipcodes$ztca5ce20) # character

# Crop US zipcodes data frame to include only CA zipcodes
#ca_zipcodes <- us_zipcodes %>% 
#  filter(ztca5ce20 > 90000,
#         ztca5ce20 < 95000)

# Check CRS (Coordinate Reference System)
#st_crs(ca_counties) # WGS 84
#st_crs(us_zipcodes) # NA
```

# Data Analysis

## Average time to take survey 

```{r}
# Check class
class(survey_data_tidy$duration_in_seconds) # character

# Convert class from character to numeric
survey_data_tidy$duration_in_seconds = as.numeric(survey_data_tidy$duration_in_seconds)

# Calculate time (in minutes) spent taking survey
mean((survey_data_tidy$duration_in_seconds)/60)  # check for outliers
median((survey_data_tidy$duration_in_seconds)/60) # positively skewed
min((survey_data_tidy$duration_in_seconds)/60) # potential outlier
max((survey_data_tidy$duration_in_seconds)/60) # potential outlier
```
