==> Rmarkdowns/3_survey_data_analysis_sr.Rmd <==
---
title: "Survey Data Exploration"
author: "Sydney Rilum"
date: "12/15/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
# Load necessary packages
library(tidyverse)
library(here)
library(janitor)
library(data.table)  # advanced pivot_table tools
library(naniar)      # replace NA tools
library(sf)          # read in shapefiles
#library(tmap)        # interactive maps
library(zipcodeR)    # for zipcode assignment by county
#library(raster)      # for mapping raster layers
#library(dplyr)       # needed ONLY if working on a windows computer
library(ggwordcloud) # create wordcloud plots
library(wordcloud)
library(patchwork)   # for stacking plots
library(kableExtra)  # for making tables

#library(conflicted)  # allows for choosing preference over conflicting functions from different packages
#conflict_prefer("select", "dplyr") # use select function from tidyverse/dplyr package > from raster package
#conflict_prefer("filter", "dplyr")
#conflict_prefer("summarize", "dplyr")
#conflict_prefer("group_by", "dplyr")
```

```{r}
# load in data (CSV created in survey_data_weighting.Rmd)
survey_data_tidy_sbv_filtered <- read.csv(here("data","survey_data_tidy_sbv_filtered.csv"))
survey_data_tidy_sbv_long_filtered <- read.csv(here("data","survey_data_tidy_sbv_long_filtered.csv"))
```

## Wording Experiment - Does word choice influence residents' support for kelp aquaculture? 

```{r}
# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming
# q6 - seaweed aquaculture
# q8 - kelp aquaculture

wording_exp <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ 1,
    answer == "Somewhat opposed" ~ 2,
    answer == "Neither opposed nor in support" ~ 3,
    answer == "Somewhat supportive" ~ 4,
    answer == "Strongly supportive" ~ 5)) %>% 
  filter(answer != "NA") 

# boxplot of wording experiment answers
ggplot(data = wording_exp, 
       aes(x = question,
           y = answer)) +
  coord_flip() +
  geom_boxplot() +
  stat_summary(fun = mean, 
               geom = "point", 
               shape = 20, 
               size = 3, 
               color = "black", 
               fill = "black") +
  labs(x = "Term",
       y = "Strongly Opposed to Strongly Supportive",
       title = "Are you supportive of or opposed to expanding ______ off the California coast?") +
  theme_light() +
  theme(legend.position = "none")

```

```{r}
# Descriptive statistics subset - wording experiment answer counts
wording_exp_counts <- wording_exp %>% 
  group_by(question) %>% 
  summarize(count1 = sum(answer == 1),
            count2 = sum(answer == 2),
            count3 = sum(answer == 3),
            count4 = sum(answer == 4),
            count5 = sum(answer == 5),
            sample_size = n())

# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming
# q6 - seaweed aquaculture
# q8 - kelp aquaculture

# summary table
wording_exp_counts %>% 
  kable(col.names = c("Question", 
                      "Strongly opposed", 
                      "Somewhat opposed", 
                      "Neither opposed nor in support",
                      "Somewhat supportive",
                      "Strongly supportive",
                      "Sample Size")) %>% 
  kable_styling()



##### NOT RIGHT ??
## How do I get means..... for each question
wording_exp_stats <- wording_exp %>% 
  group_by(question) %>% 
  summarize(mean1 = round(mean(answer == 1),3),
            mean2 = round(mean(answer == 2),3),
            mean3 = round(mean(answer == 3),3),
            mean4 = round(mean(answer == 4),3),
            mean5 = round(mean(answer == 5),3))

```



```{r non-weighted support/opposition graph}
# Combine Q2,4,6,8

# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming
# q6 - seaweed aquaculture
# q8 - kelp aquaculture

# alternative data frame
support_aq_stacked_bar <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "seaweed farming",
    question == "q4" ~ "kelp farming",
    question == "q6" ~ "seaweed aquaculture",
    question == "q8" ~ "kelp aquaculture")) %>% 
  mutate(number = case_when( 
    answer == "Strongly opposed" ~ 1,
    answer == "Somewhat opposed" ~ 2,
    answer == "Neither opposed nor in support" ~ 3,
    answer == "Somewhat supportive" ~ 4,
    answer == "Strongly supportive" ~ 5)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))

# Specify categorical order of questions and answers
support_aq_stacked_bar$answer <- factor(support_aq_stacked_bar$answer, levels = c("Strongly opposed", 
                                                                                        "Somewhat opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Somewhat supportive",
                                                                                        "Strongly supportive"))
# percent stacked barchart
ggplot(data = support_aq_stacked_bar, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  #scale_fill_manual(values =c("#2C7BB6","#2C7BB6","#FFFFBF","#FDAE61","#FDAE61")) +
  labs(x = "",
       y = "",
       title = "Are you supportive of or opposed to expanding ______ off the California coast?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))

```

```{r WEIGHTED support/oppostion graph}
# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming
# q6 - seaweed aquaculture
# q8 - kelp aquaculture

support_aq_stacked_bar_weighted <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "Seaweed Farming",
    question == "q4" ~ "Kelp Farming",
    question == "q6" ~ "Seaweed Aquaculture",
    question == "q8" ~ "Kelp Aquaculture")) %>% 
  mutate(number = case_when( 
    answer == "Strongly opposed" ~ 1,
    answer == "Somewhat opposed" ~ 2,
    answer == "Neither opposed nor in support" ~ 3,
    answer == "Somewhat supportive" ~ 4,
    answer == "Strongly supportive" ~ 5)) %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

# Specify categorical order of questions and answers
support_aq_stacked_bar_weighted$answer <- factor(support_aq_stacked_bar_weighted$answer, levels = c("Strongly opposed", 
                                                                                        "Somewhat opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Somewhat supportive",
                                                                                        "Strongly supportive"))
support_aq_stacked_bar_weighted$question <- factor(support_aq_stacked_bar_weighted$question, levels = c("Kelp Aquaculture",
                                                                                                        "Seaweed Aquaculture",
                                                                                                        "Kelp Farming",
                                                                                                        "Seaweed Farming"))

# percent stacked barchart
support_aq_stacked_bar_weighted_graph <-
  ggplot(data = support_aq_stacked_bar_weighted, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.85) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 10)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  #scale_fill_manual(values =c("#2C7BB6","#2C7BB6","#FFFFBF","#FDAE61","#FDAE61")) +
  labs(x = "",
       y = "Proportion of Responses") +
       #title = "Are you supportive of or opposed to expanding ______ off the California coast?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "none",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 12),
        axis.title.y = element_text(color = "black", size = 12),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 12, vjust = -2),
        panel.grid = element_line(color = "gray"))

support_aq_stacked_bar_weighted_graph

ggsave(filename = file.path(here("graphs","weighted_support_opposition_word_exp.png")),
       width = 8, height = 4)
```



```{r weighted support/opposition combined categories}
# alternative data frame
support_aq_stacked_bar_2_weighted <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "seaweed farming",
    question == "q4" ~ "kelp farming",
    question == "q6" ~ "seaweed aquaculture",
    question == "q8" ~ "kelp aquaculture")) %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ "Opposed",
    answer == "Somewhat opposed" ~ "Opposed",
    answer == "Neither opposed nor in support" ~ "Neither opposed nor in support",
    answer == "Somewhat supportive" ~ "Supportive",
    answer == "Strongly supportive" ~ "Supportive")) %>%
  drop_na() %>%
  mutate(number = case_when( 
    answer == "Opposed" ~ 1,
    answer == "Neither opposed nor in support" ~ 2,
    answer == "Supportive" ~ 3)) %>%
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

# Specify categorical order of questions and answers
support_aq_stacked_bar_2_weighted$answer <- factor(support_aq_stacked_bar_2_weighted$answer, levels = c("Opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Supportive"))
# percent stacked barchart
ggplot(data = support_aq_stacked_bar_2_weighted, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_fill_manual(values =c("#2C7BB6","#FFFFBF","#FDAE61")) +
  labs(x = "",
       y = "",
       title = "Are you supportive of or opposed to expanding ______ off the California coast?") +
  theme_classic() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))
```

## Support/Opposition to Aquaculture

```{r COMBINED support/opposition data to Q2}
# Combine Q2,4,6,8

# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming
# q6 - seaweed aquaculture
# q8 - kelp aquaculture

combined_support_answers <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(count = n()) %>% 
  mutate(proportion = (count/sum(count)))

# Specify categorical order of questions and answers
combined_support_answers$answer <- factor(combined_support_answers$answer, levels = c("Strongly opposed", 
                                                                                        "Somewhat opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Somewhat supportive",
                                                                                        "Strongly supportive"))

# percent stacked barchart
ggplot(data = combined_support_answers, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent) +
  #scale_fill_brewer(type = "div",
                    #palette = "RdYlGn",
                    #direction = -1) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "",
       title = "Are you supportive of or opposed to expanding \nseaweed aquaculture off the California coast?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") 

```


```{r WEIGHTED & COMBINED support/opposition data to Q2}
# Combine Q2,4,6,8

# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming
# q6 - seaweed aquaculture
# q8 - kelp aquaculture

combined_support_answers_weighted <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "combined",
    question == "q4" ~ "combined",
    question == "q6" ~ "combined",
    question == "q8" ~ "combined")) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
combined_support_answers_weighted$answer <- factor(combined_support_answers_weighted$answer, levels = c("Strongly opposed", 
                                                                                        "Somewhat opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Somewhat supportive",
                                                                                        "Strongly supportive"))

# percent stacked barchart
ggplot(data = combined_support_answers_weighted, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  #scale_fill_brewer(type = "div",
                    #palette = "RdYlGn",
                    #direction = -1) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "",
       title = "Are you supportive of or opposed to expanding \nseaweed aquaculture off the California coast?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black"),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black"),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","weighted_combined_support_opposition.png")))
```
```{r weighted combined 5 columns}

# combine data frames
weighted_support_5col <- rbind(support_aq_stacked_bar_weighted, combined_support_answers_weighted) 

# Specify categorical order of questions and answers
weighted_support_5col$answer <- factor(weighted_support_5col$answer, levels = c("Strongly opposed", 
                                                                                        "Somewhat opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Somewhat supportive",
                                                                                        "Strongly supportive"))
weighted_support_5col$question <- factor(weighted_support_5col$question, levels = c("kelp aquaculture", 
                                                                                        "kelp farming", 
                                                                                        "seaweed aquaculture",
                                                                                        "seaweed farming",
                                                                                        "combined"))
# percent stacked barchart
ggplot(data = weighted_support_5col, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  #scale_fill_manual(values =c("#2C7BB6","#2C7BB6","#FFFFBF","#FDAE61","#FDAE61")) +
  labs(x = "",
       y = "weighted proportion of repsonses",
       title = "Are you supportive of or opposed to expanding ______ off the California coast?") +
  theme_classic() +
  theme(legend.title = element_blank(),
        legend.position = "right") 

#ggsave("weighted_support_5col.png", width = 8, height = 4)

```

```{r WEIGHTED & COMBINED 2 COLUMN support/opposition data}
# Combine Q2&4 vs. Q6&8

# Are you supportive of or opposed to expanding ______ off the California coast?
# q2 - seaweed farming
# q4 - kelp farming

# q6 - seaweed aquaculture
# q8 - kelp aquaculture

combined_support_answers_weighted_2col <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "Farming",
    question == "q4" ~ "Farming",
    question == "q6" ~ "Aquaculture",
    question == "q8" ~ "Aquaculture")) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
combined_support_answers_weighted_2col$answer <- factor(combined_support_answers_weighted_2col$answer, levels = c("Strongly opposed", 
                                                                                        "Somewhat opposed", 
                                                                                        "Neither opposed nor in support",
                                                                                        "Somewhat supportive",
                                                                                        "Strongly supportive"))

# percent stacked barchart
combined_support_answers_weighted_2col_graph <- 
  ggplot(data = combined_support_answers_weighted_2col, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.75) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  #scale_fill_brewer(type = "div",
                    #palette = "RdYlGn",
                    #direction = -1) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "") +
       #title = "Influence of Terminology on Support") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "none",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 12),
        #axis.title.y = element_text(color = "black", size = 12)
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 12, vjust = -2),
        panel.grid = element_line(color = "gray"),
        plot.title = element_text(hjust = 0.5))

combined_support_answers_weighted_2col_graph

ggsave(filename = file.path(here("graphs","weighted_combined_support_opposition_2col.png")))
```

```{r, support/opposition PATCHWORK}
# stack  plots using `patchwork` 
support_aq_stacked_bar_weighted_graph + plot_spacer() + combined_support_answers_weighted_2col_graph +
  plot_annotation(title = "Influence of Terminology on Support", tag_levels = "A") +
  theme(plot.title = element_text(size = 15)) +
  plot_layout(widths = c(5, 0.01, 3))

ggsave(filename = file.path(here("graphs","influence_terms_on_support_PATCHWORK.png")),
       width = 10, height = 6)

```


### Word Clouds 

```{r}
# Word Cloud data cleaning

# convert each string of words to lowercase
survey_data_tidy$q3 <- tolower(survey_data_tidy$q3)  # seaweed farming
survey_data_tidy$q5 <- tolower(survey_data_tidy$q5)  # kelp farming
survey_data_tidy$q7 <- tolower(survey_data_tidy$q7)  # seaweed aquaculture
survey_data_tidy$q9 <- tolower(survey_data_tidy$q9)  # seaweed farming
```

```{r}
# q3 - seaweed farming
seaweed_farming <- survey_data_tidy %>% 
  select(q3) %>%
  mutate(q3 = str_trim(q3)) %>%  # remove excess whitespace before/after word
  mutate(q3 = str_replace(q3, pattern = "goodl", replacement = "good")) %>% 
  mutate(q3 = str_replace(q3, pattern = "auaculture", replacement = "auqaculture")) %>% 
  mutate(q3 = str_replace(q3, pattern = "sustainability", replacement = "sustainable")) %>% 
  filter(q3 != "idk") %>% 
  filter(q3 != "huh") %>% 
  filter(q3 != "none") %>% 
  filter(q3 != "nothing") %>% 
  count(q3) %>%                  # count repeating words
  arrange(-n)                    # arrange word counts from largest to smallest

wordcloud_seaweed_farming <- ggplot(data = seaweed_farming, 
                                    aes(label = q3)) +
  geom_text_wordcloud(aes(color = n, size = n), 
                      shape = "diamond") +
  scale_size_area(max_size = 25) +
  scale_color_gradientn(colors = c("burlywood4", "green")) +
  theme_minimal() +
  labs(title = "seaweed farming") 
  #theme(axis.line.y = element_line(color="black", size = 0.1))

# alternative word cloud
seaweed_farming_test <- seaweed_farming %>% 
  slice(-1) # remove first row of df

wordcloud_seaweed_farming_2 <- wordcloud(words = seaweed_farming_test$q3, 
                                         freq = seaweed_farming_test$n,
                                         min.freq = 1,
                                         random.order = FALSE,
                                         colors = brewer.pal(8, "Dark2")) 
```

```{r}
# q5 - kelp farming
kelp_farming <- survey_data_tidy %>% 
  select(q5) %>%
  mutate(q5 = str_trim(q5)) %>%  # remove excess whitespace before/after word
  mutate(q5 = str_replace(q5, pattern = "icean", replacement = "ocean")) %>% 
  filter(q5 != "i") %>% 
  filter(q5 != "idk") %>% 
  filter(q5 != "what") %>% 
  filter(q5 != "test") %>% 
  filter(q5 != "testpanel") %>%
  count(q5) %>%                  # count repeating words
  arrange(-n)                    # arrange word counts from largest to smallest


wordcloud_kelp_farming <- ggplot(data = kelp_farming, 
                                    aes(label = q5)) +
  geom_text_wordcloud(aes(color = n, size = n), 
                      shape = "diamond") +
  scale_size_area(max_size = 25) +
  scale_color_gradientn(colors = c("darkgreen", "green")) +
  theme_minimal() +
  labs(title = "kelp farming") +
  theme(plot.title.position = "panel")
  #theme(axis.line.x = element_line(color="black", size = 0.1),
        #axis.line.y = element_line(color="black", size = 0.1))

# alternative word cloud
kelp_farming_test <- kelp_farming %>% 
  slice(-1) # remove first row of df

wordcloud_kelp_farming_2 <- wordcloud(words = kelp_farming_test$q5, 
                                      freq = kelp_farming_test$n,
                                      min.freq = 1,
                                      random.order = FALSE,
                                      colors = brewer.pal(8, "Dark2"))
```

```{r}
# q7 - seaweed aquaculture
seaweed_aquaculture <- survey_data_tidy %>% 
  select(q7) %>%
  mutate(q7 = str_trim(q7)) %>%  # remove excess whitespace before/after word
  mutate(q7 = str_replace(q7, pattern = "enviironmental", replacement = "environment")) %>% 
  mutate(q7 = str_replace(q7, pattern = "enviroment", replacement = "environment")) %>% 
  mutate(q7 = str_replace(q7, pattern = "environments", replacement = "environment")) %>% 
  mutate(q7 = str_replace(q7, pattern = "harvesting", replacement = "harvest")) %>% 
  filter(q7 != "none") %>% 
  filter(q7 != "nothing") %>% 
  count(q7) %>%                  # count repeating words
  arrange(-n)                    # arrange word counts from largest to smallest


wordcloud_seaweed_aquaculture <- ggplot(data = seaweed_aquaculture, 
                                    aes(label = q7)) +
  geom_text_wordcloud(aes(color = n, size = n), 
                      shape = "diamond") +
  scale_size_area(max_size = 25) +
  scale_color_gradientn(colors = c("springgreen4", "green")) +
  theme_minimal() +
  labs(title = "seaweed aquaculture")

# alternative word cloud
seaweed_aquaculture_test <- seaweed_aquaculture %>% 
  slice(-1) # remove first row of df

wordcloud_seaweed_aquaculture_2 <- wordcloud(words = seaweed_aquaculture_test$q7, 
                                            freq = seaweed_aquaculture_test$n,
                                            min.freq = 1,
                                            random.order = FALSE,
                                            colors = brewer.pal(8, "Dark2"))
```

```{r}
# q9 - kelp aquaculture
kelp_aquaculture <- survey_data_tidy %>% 
  select(q9) %>%
  mutate(q9 = str_trim(q9)) %>%  # remove excess whitespace before/after word
  mutate(q9 = str_replace(q9, pattern = "aquaculter", replacement = "aquaculture")) %>% 
  mutate(q9 = str_replace(q9, pattern = "environment", replacement = "environmental")) %>% 
  mutate(q9 = str_replace(q9, pattern = "environmentalal", replacement = "environmental")) %>%   
  mutate(q9 = str_replace(q9, pattern = "sustainable", replacement = "sustainability")) %>%   
  filter(q9 != "huh") %>% 
  filter(q9 != "noidea") %>% 
  filter(q9 != "none") %>% 
  filter(q9 != "nope") %>% 
  filter(q9 != "unsure") %>% 
  count(q9) %>%                  # count repeating words
  arrange(-n)                    # arrange word counts from largest to smallest


wordcloud_kelp_aquaculture <- ggplot(data = kelp_aquaculture, 
                                    aes(label = q9)) +
  geom_text_wordcloud(aes(color = n, size = n), 
                      shape = "diamond") +
  scale_size_area(max_size = 25) +
  scale_color_gradientn(colors = c("yellow4", "green")) +
  theme_minimal() +
  labs(title = "kelp aquaculture") 
  #theme(axis.line.x = element_line(color="black", size = 0.1))

# alternative word cloud
kelp_aquaculture_test <- kelp_aquaculture %>% 
  slice(-1) # remove first row of df

wordcloud_kelp_aquaculture_2 <- wordcloud(words = kelp_aquaculture_test$q9, 
                                         freq = kelp_aquaculture_test$n,
                                         min.freq = 1,
                                         random.order = FALSE,
                                         colors = brewer.pal(8, "Dark2"))
```

```{r}
# stack word cloud plots using `patchwork` 
wordclouds <- ((wordcloud_seaweed_farming + wordcloud_seaweed_aquaculture) /
                 (wordcloud_kelp_farming + wordcloud_kelp_aquaculture))

wordclouds +
  plot_annotation(title = "When you think of _____  _____, what word comes to mind?") &
  theme(plot.title = element_text(size = 15)) 

```



### Ranking terms by positivity 

```{r, unweighted ranked terms}
# Q11_1 - Q11_8 (Aquaculture is referred to by many names. Please rank the following names based on which you view most positively, with 1 being most positive.)

# q11_1 - ocean farming
# q11_2 - seaweed farming
# q11_3 - kelp farming
# q11_4 - ocean aquaculture
# q11_5 - seaweed aquaculture
# q11_6 - kelp aquaculture
# q11_7 - regenerative aquaculture
# q11_8 - seaweed mariculture

# make new subset of data (WITH ALL SURVEY RESPONSES)
wording_exp_rank <- survey_data_tidy %>% 
  select(q11_1:q11_8) %>% 
  mutate_if(is.character,as.numeric) %>% 
  summarize_all(mean, na.rm = TRUE) # calculate mean rank score for each answer choice

# transpose dataframe
wording_exp_rank <- t(wording_exp_rank) %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "term") %>% 
  mutate(term = case_when(
    term == "q11_1" ~ "ocean farming",
    term == "q11_2" ~ "seaweed farming",
    term == "q11_3" ~ "kelp farming",
    term == "q11_4" ~ "ocean aquaculture",
    term == "q11_5" ~ "seaweed aquaculture",
    term == "q11_6" ~ "kelp aquaculture",
    term == "q11_7" ~ "regenerative aquaculture",
    term == "q11_8" ~ "seaweed mariculture")) %>% 
  mutate(term = fct_reorder(term, V1, .desc = TRUE))  # order topics by count, smallest (most interested) to largest (least interested)

ggplot(data = wording_exp_rank, aes(x = term, y = V1)) +
  geom_col(fill = "skyblue2") +
  coord_flip() +
  theme_minimal() +
  labs(x = "",
       y = "Average Ranked Score (from 1 - 6, with 1 being most positive)",
       title = "Terminology Ranked by Most Positive Connotation")

```

```{r, weighted ranked terms}
# Q11_1 - Q11_8 (Aquaculture is referred to by many names. Please rank the following names based on which you view most positively, with 1 being most positive.)

# q11_1 - ocean farming
# q11_2 - seaweed farming
# q11_3 - kelp farming
# q11_4 - ocean aquaculture
# q11_5 - seaweed aquaculture
# q11_6 - kelp aquaculture
# q11_7 - regenerative aquaculture
# q11_8 - seaweed mariculture


# make new subset of data USING SBV data ONLY
wording_exp_rank_2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q11_1"|
           question == "q11_2"|
           question == "q11_3"|
           question == "q11_4"|
           question == "q11_5"|
           question == "q11_6"|
           question == "q11_7"|
           question == "q11_8") %>% 
  mutate(question = case_when(
    question == "q11_1" ~ "ocean farming",
    question == "q11_2" ~ "seaweed farming",
    question == "q11_3" ~ "kelp farming",
    question == "q11_4" ~ "ocean aquaculture",
    question == "q11_5" ~ "seaweed aquaculture",
    question == "q11_6" ~ "kelp aquaculture",
    question == "q11_7" ~ "regenerative aquaculture",
    question == "q11_8" ~ "seaweed mariculture")) %>% 
  drop_na() %>% 
  mutate(weight = as.numeric(weight)) %>% 
  group_by(question) #%>%
  #summarize(mean = sum(weight, na.rm = TRUE))


 # summarize(mean = round(mean(weight, na.rm = TRUE),2), # calculate mean rank score for each answer choice
 #           sd = round(sd(weight, na.rm = TRUE),2)) %>% 
 # mutate(question = fct_reorder(question, mean, .desc = TRUE))  # order topics by count, smallest (most interested) to largest (least interested)


ggplot(data = wording_exp_rank_2, aes(x = mean, 
                                      y = question,
                                      fill = mean)) +
  geom_bar(stat = "identity",
           fill = "skyblue2") +
  geom_text(aes(label = mean),
            hjust = 1.5) +
  labs(y = "",
       x = "Average Ranked Score (from 1 - 6, with 1 being most positive)",
       title = "Terminology Ranked by Most Positive Connotation") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "none") 

#ggsave("rank_terms.png")
```


## Knowledge 

```{r}
# (Q10) - How familiar are you with aquaculture?

familiarity <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q10") %>% 
  group_by(answer) %>% 
  summarize(weight_count = round(sum(weight),)) %>% 
  mutate(proportion = (weight_count/sum(weight_count))) 


# Specify categorical order of answers
familiarity$answer <- factor(familiarity$answer, levels = c("Never heard of it",
                                                            "Heard of it, but don't know details",
                                                            "Basic understanding",
                                                            "Very familiar"))

ggplot(data = familiarity, aes(x = proportion,
                               y = answer,
                               fill = answer)) +
  geom_bar(stat = "identity",
           fill = "#ABD9E9") + 
  #"#2C7BB6","#ABD9E9"
  #geom_text(aes(label = proportion), hjust = 1.5) +
  scale_x_continuous(labels = scales::label_percent(accuracy = 1L),
                     expand = expansion(mult = c(0, 0))) +
  scale_y_discrete(labels= function(y)str_wrap(y, width = 12)) +
  labs(x = "Proportion of Responses",
       y = "",
       title = "How familiar are you with aquaculture?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "none",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 12),
        axis.title.x = element_text(color = "black", size = 12),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 12),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","weighted_familiarity.png")))
```



## Perception of Aquaculture vs. Seaweed Aquaculture 

```{r}
## Comparing perceptions

# (Q12) - In general, how positive or negative is your view of aquaculture?
# (Q13) - In general, how positive or negative is your view of seaweed aquaculture?

perception_aq_boxplot <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q12" | question == "q13") %>% 
  mutate(question = case_when( 
    question == "q12" ~ "aquaculture",
    question == "q13" ~ "seaweed aquaculture")) %>% 
  mutate(answer = case_when( 
    answer == "Very Negative" ~ -2,
    answer == "Negative" ~ -1,
    answer == "Neutral" ~ 0,
    answer == "Positive" ~ 1,
    answer == "Very Positive" ~ 2,
    answer == "Unsure/Not familiar" ~ -999)) %>%
  replace_with_na(replace = list(answer = -999))
  
# boxplot  
ggplot(data = perception_aq_boxplot, 
       aes(x = question,
           y = answer)) +
  coord_flip() +
  geom_boxplot(color = c("darkblue","springgreen4")) +
  stat_summary(fun = mean, 
               geom = "point", 
               shape = 20, 
               size = 3, 
               color = "black", 
               fill = "black") +
  labs(x = "",
       y = "Very Negative to Very Positive",
       title = "In general, how positive or negative is your view of ______ ?") +
  theme_light() +
  theme(legend.position = "none")

```

```{r, unweighted perception}
# alternative data frame
perception_aq_stacked_bar <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q12" | question == "q13") %>% 
  mutate(question = case_when( 
    question == "q12" ~ "aquaculture",
    question == "q13" ~ "seaweed aquaculture")) %>% 
  mutate(number = case_when( 
    answer == "Very Negative" ~ 1,
    answer == "Negative" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Positive" ~ 4,
    answer == "Very Positive" ~ 5,
    answer == "Unsure/Not familiar" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of questions and answers
perception_aq_stacked_bar$answer <- factor(perception_aq_stacked_bar$answer, levels = c("Unsure/Not familiar", 
                                                                                        "Very Negative", 
                                                                                        "Negative",
                                                                                        "Neutral",
                                                                                        "Positive",
                                                                                        "Very Positive"))
perception_aq_stacked_bar$question <- factor(perception_aq_stacked_bar$question, levels = c("aquaculture",
                                                                                            "seaweed aquaculture"))

# percent stacked barchart
ggplot(data = perception_aq_stacked_bar, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "In general, how positive or negative is your view of ______ ?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))

```

```{r}
# Descriptive Statistics
perception_aq_stats <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q12" |
           question == "q13") %>% 
  mutate(answer = case_when( 
    answer == "Very Negative" ~ -2,
    answer == "Negative" ~ -1,
    answer == "Neutral" ~ 0,
    answer == "Positive" ~ 1,
    answer == "Very Positive" ~ 2,
    answer == "Unsure/Not familiar" ~ -999)) %>%
  replace_with_na(replace = list(answer = -999)) %>% 
  drop_na() %>% 
  group_by(question) %>% 
  summarize(mean = mean(answer),
            median = median(answer),
            sd = sd(answer))
  
```

```{r, WEIGHTED perception}
# alternative data frame
perception_aq_stacked_bar_weighted <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q12" | question == "q13") %>% 
  mutate(question = case_when( 
    question == "q12" ~ "Aquaculture",
    question == "q13" ~ "Seaweed Aquaculture")) %>% 
  mutate(number = case_when( 
    answer == "Very Negative" ~ 1,
    answer == "Negative" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Positive" ~ 4,
    answer == "Very Positive" ~ 5,
    answer == "Unsure/Not familiar" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
perception_aq_stacked_bar_weighted$answer <- factor(perception_aq_stacked_bar_weighted$answer, levels = c("Unsure/Not familiar", 
                                                                                        "Very Negative", 
                                                                                        "Negative",
                                                                                        "Neutral",
                                                                                        "Positive",
                                                                                        "Very Positive"))
perception_aq_stacked_bar_weighted$question <- factor(perception_aq_stacked_bar_weighted$question, levels = c("Aquaculture",
                                                                                            "Seaweed Aquaculture"))

# percent stacked barchart
ggplot(data = perception_aq_stacked_bar_weighted, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 10)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "In general, how positive or negative is your view of ______ ?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 12),
        axis.title.y = element_text(color = "black", size = 12),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 12),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","weighted_perception.png")))

```




## Environmental Impacts

```{r}
# Environmental benefits/concerns (Q23)

# Q23_1 - Q23_7
# To what extent do you agree with each of the following statements?

environ_impacts <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q23_1" | 
           question == "q23_2" |
           question == "q23_3" |
           question == "q23_4" |
           question == "q23_5" |
           question == "q23_6" |
           question == "q23_7") %>% 
  mutate(question = case_when( 
    question == "q23_1" ~ "...has negative impacts on marine ecosystems.",
    question == "q23_2" ~ "...improves water quality nearby.",
    question == "q23_3" ~ "...infrastructure poses a risk to marine life.",
    question == "q23_4" ~ "...infrastructure has positive impacts on marine ecosystems.",
    question == "q23_5" ~ "...has positive impacts on marine ecosystems.",
    question == "q23_6" ~ "...causes bad smells nearby.",
    question == "q23_7" ~ "...is visually appealing.")) %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  #replace_with_na(replace = list(number = -999)) %>% 
  drop_na() %>%                                            # drop NAs (counts of those who skipped question)
  #mutate_all(na_if, "") %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count))*100) %>% 
  ungroup() 

# Specify categorical order of answers
environ_impacts$answer <- factor(environ_impacts$answer, levels = c("Strongly Disagree", 
                                                                    "Disagree", 
                                                                    "Neutral",
                                                                    "Agree",
                                                                    "Strongly Agree",
                                                                    "Don't know"))

# percent stacked barchart
ggplot(data = environ_impacts, 
       aes(fill = forcats::fct_rev(answer), 
           x = proportion,
           y = question)) +
  geom_bar(position = "fill",
           stat = "identity") +
  scale_x_continuous(labels = scales::percent) +
  scale_y_discrete(position = "right") +
  scale_fill_brewer(type = "div",
                    palette = "RdYlBu",
                    direction = +1) +
  labs(x = "",
       y = "",
       title = "Seaweed Aquaculture...") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "bottom") +
  guides(fill=guide_legend(nrow=1,byrow=TRUE))

```


```{r}
# environmental benefits/concerns (Q24)

# create new data frame
environ_impacts_2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q24_1" | 
           question == "q24_2" |
           question == "q24_4") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
environ_impacts_2$answer <- factor(environ_impacts_2$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = environ_impacts_2, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))
```


## Economic Impacts

```{r}
# economic benefits/concerns (Q25)

# create new data frame
econ_impacts <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q25_1" | 
           question == "q25_2" |
           question == "q25_3" |
           question == "q25_4" |
           question == "q25_5") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
econ_impacts$answer <- factor(econ_impacts$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = econ_impacts, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "economic impacts: seaweed aquaculture...") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))

```


```{r}
# Seaweed Aquaculture risks (Q26)

# create new data frame
swaq_risks <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q26_1" | 
           question == "q26_2" |
           question == "q26_3") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
swaq_risks$answer <- factor(swaq_risks$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = swaq_risks, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))
```

```{r}
# Concerns (Q27)

# create new data frame
swaq_concerns <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q27_1" | 
           question == "q27_2" |
           question == "q27_3" |
           question == "q27_4" |
           question == "q27_5" |
           question == "q27_6" |
           question == "q27_7" |
           question == "q27_8") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
swaq_concerns$answer <- factor(swaq_concerns$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = swaq_concerns, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))
```


```{r}
# Seaweed Aquaculture Interest (Q28)

# create new data frame
swaq_interest <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q28_1" | 
           question == "q28_2" |
           question == "q28_3") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
swaq_interest$answer <- factor(swaq_interest$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = swaq_interest, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))
```

## graphs for faculty review


```{r}
## graph 1

# create new data frame
graph_1 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q23_1" | 
           question == "q23_3" |
           question == "q26_3") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
graph_1$answer <- factor(graph_1$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = graph_1, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))




```

```{r}
# graph 2
graph_2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q24_4" | 
           question == "q25_1" |
           question == "q25_4") %>% 
  mutate(number = case_when( 
    answer == "Strongly Disagree" ~ 1,
    answer == "Disagree" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Agree" ~ 4,
    answer == "Strongly Agree" ~ 5,
    answer == "Don't know" ~ 6)) %>%
  drop_na() %>% 
  group_by(question, answer, number) %>% 
  summarize(count = n()) %>% 
  ungroup() %>% 
  group_by(question) %>% 
  mutate(proportion = (count / sum(count)))


# Specify categorical order of answers
graph_2$answer <- factor(graph_2$answer, levels = c("Don't know",
                                                              "Strongly Disagree", 
                                                              "Disagree", 
                                                              "Neutral",
                                                              "Agree",
                                                              "Strongly Agree"))

# percent stacked barchart
ggplot(data = graph_2, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(x = "",
       y = "",
       title = "") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right") #+
  #guides(color = guide_legend(reverse=TRUE),
  #       fill = guide_legend(nrow=1, byrow=TRUE))
```

## used in final report!! (for environmental impacts)

```{r}
# trust in scientists (Q27_2)
scientists_trust <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q27_2" | 
           question == "q27_3") %>% 
  mutate(question = case_when(
    question == "q27_2" ~ "I trust scientists who study how we use the environment.",
    question == "q27_3" ~ "I am concerned about climate change."
  )) %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = round(sum(weight),)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
scientists_trust$answer <- factor(scientists_trust$answer, levels = c("Don't know", "Strongly Disagree", "Disagree","Neutral","Agree","Strongly Agree"))

# percent stacked barchart
ggplot(data = scientists_trust, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 25)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(y = "Proportion of Responses",
       x = "",
       title = "To what extent do you agree with each of the following statements?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","scientists_trust.png")))
```

```{r}
# envi impacts
envi_impacts <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q23_1" | 
           question == "q23_3" |
           question == "q26_2" |
           question == "q26_3") %>% 
  mutate(question = case_when(
    question == "q23_1" ~ "Seaweed aquaculture has negative impacts on marine ecosystems.",
    question == "q23_3" ~ "Seaweed aquaculture infrastructure poses a risk to marine life.",
    question == "q26_2" ~ "Seaweed aquaculture has fewer risks than shellfish and fish aquaculture.",
    question == "q26_3" ~ "The benefits of seaweed aquaculture outweigh the potential risks."
  )) %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = round(sum(weight),)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
envi_impacts$answer <- factor(envi_impacts$answer, levels = c("Don't know", "Strongly Disagree", "Disagree","Neutral","Agree","Strongly Agree"))

envi_impacts$question <- factor(envi_impacts$question, levels = c("Seaweed aquaculture has negative impacts on marine ecosystems.","Seaweed aquaculture infrastructure poses a risk to marine life.","Seaweed aquaculture has fewer risks than shellfish and fish aquaculture.","The benefits of seaweed aquaculture outweigh the potential risks."))

# percent stacked barchart
ggplot(data = envi_impacts, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 25)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(y = "Proportion of Responses",
       x = "",
       title = "To what extent do you agree with each of the following statements?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","envi_impacts.png")),
       width = 10, height = 7)
```

```{r}
# econ impacts
econ_impacts <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q24_4" | 
           question == "q25_1" |
           question == "q25_2" |
           question == "q25_4" |
           question == "q25_5") %>% 
  mutate(question = case_when(
    question == "q24_4" ~ "Aquaculture will be an important food supply as climate change progresses.",
    question == "q25_1" ~ "Seaweed aquaculture has a positive impact on the local economy",
    question == "q25_2" ~ "Seaweed aquaculture lowers the price of seafood.",
    question == "q25_4" ~ "Seaweed aquaculture interferes with commercial and recreational fishing.",
    question == "q25_5" ~ "Seaweed aquaculture interferes with tourism/recreation.")) %>%
  drop_na(answer) %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = round(sum(weight),)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
econ_impacts$answer <- factor(econ_impacts$answer, levels = c("Don't know", "Strongly Disagree", "Disagree","Neutral","Agree","Strongly Agree"))

econ_impacts$question <- factor(econ_impacts$question, levels = c("Aquaculture will be an important food supply as climate change progresses.","Seaweed aquaculture has a positive impact on the local economy","Seaweed aquaculture lowers the price of seafood.", "Seaweed aquaculture interferes with commercial and recreational fishing.","Seaweed aquaculture interferes with tourism/recreation."))

# percent stacked barchart
ggplot(data = econ_impacts, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = question)) +
  geom_bar(position = "stack",
           stat = "identity") +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 25)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C","#C0C0C0")) +
  labs(y = "Proportion of Responses",
       x = "",
       title = "To what extent do you agree with each of the following statements?") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","econ_impacts.png")),
       width = 12, height = 7)
```











## Demographics 

```{r}
## Exploratory Plots

# Birth Year (Q31)
ggplot(data = survey_data_tidy, aes(y = q31)) +
  geom_bar()
### need to remove SB and NJ words from years

# Gender Identity (Q32)
ggplot(data = survey_data_tidy, aes(x = q32)) +
  geom_bar()

# Race/ethnicity (Q33)
ggplot(data = survey_data_tidy, aes(y = q33)) +
  geom_bar()
### how to deal with people selecting more than one option ????

# Highest level of education (Q34)
ggplot(data = survey_data_tidy, aes(y = q34)) +
  geom_bar()

# Annual household income (Q35)
ggplot(data = survey_data_tidy, aes(y = q35)) +
  geom_bar()

# Political Affiliation (Q38)
ggplot(data = survey_data_tidy, aes(y = q38)) +
  geom_bar()


```


# Demographic Analysis for Qualtrics survey (350 weighted responses)

## Political Affiliation
```{r}
# split up support/opposed responses by political affiliation
polit_affil <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "q38", "answer", "weight") %>% 
  mutate(q38 = case_when( 
    q38 == "Indpendent" ~ "Independent",
    TRUE ~ q38)) %>% 
  filter(answer != "") %>% 
  #drop_na() %>% 
  group_by(q38, answer) %>% 
  #summarize(count = n()) %>% 
  #mutate(proportion = (count / sum(count)))
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
polit_affil$answer <- factor(polit_affil$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

# percent stacked barchart
ggplot(data = polit_affil, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = q38)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  #scale_x_discrete(labels= function(x) str_wrap(x, width = 10)) +
  #scale_fill_brewer(type = "div",
                    #palette = "RdYlGn",
                    #direction = -1) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Political Affiliation and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_polit_affil.png")))
```
## Sex
```{r}
# split up support/opposed responses by sex
sex <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  mutate(q32 = case_when( 
    q32 == "Man" ~ "Male",
    q32 == "Woman" ~ "Female")) %>% 
  select("question", "q32", "answer", "weight") %>% 
  filter(answer != "") %>% 
  #drop_na() %>% 
  group_by(q32, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
sex$answer <- factor(sex$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

# percent stacked barchart
ggplot(data = sex, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = q32)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  #scale_x_discrete(labels= function(x) str_wrap(x, width = 10)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Sex and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_sex.png")))
```


## Education
```{r}
# split up support/opposed responses by education
education <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "q34", "answer", "weight") %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(q34, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
education$answer <- factor(education$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

education$q34 <- factor(education$q34, levels = c("Some High School", "High School or equivalent GED", "Technical or occupational certificate", "Some college coursework completed", "Associate degree","Bachelor's degree","Master's degree", "PhD"))

# percent stacked barchart
ggplot(data = education, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = q34)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 24)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Education and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10, angle = 45, hjust = 1),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_education.png")))
```


## Income
```{r}
# split up support/opposed responses by income
income <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "q35", "answer", "weight") %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(q35, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
income$answer <- factor(income$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

income$q35 <- factor(income$q35, levels = c("< $25K", "$25K - $50K", "$50K - $75K","$75K - $100K","$100K - $125K", "$125K - $150K", "> $150K", "Prefer not to state"))

# percent stacked barchart
ggplot(data = income, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = q35)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Household Income and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10, angle = 45, hjust = 1),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_income.png")))
```


## Relationship to Aquaculture Industry
```{r}
# split up support/opposed responses by relation to aquaculture industry
aq_industry <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "q40", "answer", "weight") %>% 
  filter(answer != "") %>% 
  drop_na() %>%
  separate(col = "q40", sep = ",", into = c("q40_1","q40_2","q40_3")) %>% 
  group_by(q40_1, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
aq_industry$answer <- factor(aq_industry$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

aq_industry$q40_1 <- factor(aq_industry$q40_1, levels = c("Employee", "Supplier", "Customer", "Competitor", "Government/Regulator", "Researcher", "Shareholder", "Other", "None"))

# percent stacked barchart
ggplot(data = aq_industry, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = q40_1)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 10)) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Relation to Aquaculture Industry and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10, angle = 45, hjust = 1),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_aq_industry.png")))
```

## County
```{r}
# split up support/opposed responses by county
county <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "county", "answer", "weight") %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(county, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

# Specify categorical order of questions and answers
county$answer <- factor(county$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

# percent stacked barchart
ggplot(data = county, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = county)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "County of Residence and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_county.png")))
```


## Age
```{r}
# split up support/opposed responses by age
age <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "age_group", "answer", "weight") %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  group_by(age_group, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))


# Specify categorical order of questions and answers
age$answer <- factor(age$answer, levels = c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support","Somewhat supportive","Strongly supportive"))

# percent stacked barchart
ggplot(data = age, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = age_group)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Age and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10, angle = 45, hjust = 1),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_age.png")))
```

## Occupation Sector
```{r}
# split up support/opposed responses by occupation sector
occupation <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  select("question", "q39", "answer", "weight") %>% 
  mutate(answer = fct_relevel(answer, c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support", "Somewhat supportive", "Strongly supportive"))) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  separate(col = "q39", sep = ",", into = c("q39_1","q39_2","q39_3")) %>% 
  group_by(q39_1, answer) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

occupation$q39_1 <- factor(occupation$q39_1, levels = c("government", "non-profit", "education","commercial fisher","healthcare", "service industry", "other"))

# percent stacked barchart
ggplot(data = occupation, 
       aes(fill = forcats::fct_rev(answer),
           y = proportion,
           x = q39_1)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF","#FDAE61","#D7191C")) +
  labs(x = "",
       y = "Proportion of Responses",
       title = "Occupation Sector and Support for Seaweed Aquaculture") +
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10, angle = 45, hjust = 1),
        panel.grid = element_line(color = "gray"))

ggsave(filename = file.path(here("graphs","demo_occupation_sector.png")))
```


## Days by the ocean
```{r}
# split up support/opposed responses by # days spent by ocean
ocean_time <- survey_data_tidy_sbv_filtered %>% 
  select(response_id, q15, q2, q4, q6, q8, weight) %>% 
  #change non-responses to '0 days' because but one non-response corresponds to 'no' response to q14 'do you spend free time by the ocean?'
  mutate(q15= case_when(
    q15 == "" ~ "0 days", 
    TRUE ~ q15)) %>% 
  #combine support metrics from multiple columns into one columns
  unite("support", q2:q8, sep ="", na.rm = TRUE) %>% 
  #filter out non-response for q2-q8
  filter(support != "") %>% 
  mutate(support = fct_relevel(support, c("Strongly opposed", "Somewhat opposed", "Neither opposed nor in support", "Somewhat supportive", "Strongly supportive"))) %>% 
  mutate(q15 = fct_explicit_na(q15, na_level = "0 days")) %>% 
  mutate(q15 = fct_relevel(q15, c("0 days", "1-5 days", "5-10 days", "10-15 days", "15+ days"))) %>% 
  group_by(q15, support) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

ggplot(data = ocean_time, 
       aes(fill = forcats::fct_rev(support),
           y = proportion,
           x = q15)) +
  geom_bar(position = "stack",
           stat = "identity",
           width = 0.5) +
  scale_y_continuous(labels = scales::percent,
                     expand = expansion(mult = c(0, 0))) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 10))+
  scale_fill_manual(values =c("#2C7BB6","#ABD9E9","#FFFFBF", "#FDAE61","#D7191C","#999999")) +
  labs(y = "Proportion of Responses",
       x = "",
       title = "Monthly Ocean Visitation and Support for Seaweed Aquaculture") +
       #fill = str_wrap("Support/Opposition", width = 15))+
  theme_minimal() +
  theme(legend.title = element_blank(),
        legend.position = "right",
        axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black", size = 10),
        axis.title.y = element_text(color = "black", size = 10),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black", size = 10, angle = 45, hjust = 1),
        panel.grid = element_line(color = "gray"))
  
ggsave(filename = file.path(here("graphs","demo_days_by_ocean.png")))
```



```{r}
# opposed residents
opposed_residents <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  filter(answer == "Strongly opposed" |
           answer == "Somewhat opposed") %>% 
  group_by(county, q32) %>% 
  summarize(count = n()) %>% 
  mutate(proportion = (count / sum(count)))

# supportive residents
supportive_residents <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  filter(answer == "Strongly supporitve" |
           answer == "Somewhat supportive") %>% 
  group_by(county, q32) %>% 
  summarize(count = n()) %>% 
  mutate(proportion = (count / sum(count)))
  
# neutral residents
neutral_residents <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  drop_na() %>% 
  filter(answer == "Neither opposed nor in support") %>% 
  group_by(county, q32) %>% 
  summarize(count = n()) %>% 
  mutate(proportion = (count / sum(count)))
```


## Communication Questions 

```{r}
# Q29_1 - Q29_6 (Rank the aquaculture related topics in order of interest in learning more, with 1 being most interested)

# q29_1 - policy
# q29_2 - permitting
# q29_3 - environmental impacts
# q29_4 - economic impacts
# q29_5 - social & cultural impacts
# q29_6 - research

# make new subset of data
comm_data <- survey_data_tidy_sbv_filtered %>% 
  select("q29_1", "q29_2", "q29_3", "q29_4", "q29_5", "q29_6", "weight") %>% 
  mutate_if(is.character,as.numeric) %>% 
  summarize_all(mean, na.rm = TRUE) # calculate mean rank score for each answer choice
  #column_to_rownames(q29_1, "policy")

# transpose dataframe
comm_data <- t(comm_data) %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "aquaculture_topics") %>% 
  mutate(aquaculture_topics = case_when(
    aquaculture_topics == "q29_1" ~ "policy",
    aquaculture_topics == "q29_2" ~ "permitting",
    aquaculture_topics == "q29_3" ~ "environmental impacts",
    aquaculture_topics == "q29_4" ~ "economic impacts",
    aquaculture_topics == "q29_5" ~ "social & cultural impacts",
    aquaculture_topics == "q29_6" ~ "research")) %>% 
  mutate(aquaculture_topics = fct_reorder(aquaculture_topics, V1, .desc = TRUE))  # order topics by count, smallest (most interested) to largest (least interested)

ggplot(data = comm_data, aes(x = aquaculture_topics, y = V1)) +
  geom_col(fill = "skyblue4") +
  coord_flip() +
  theme_minimal() +
  labs(x = "",
       y = "Average Ranked Score (from 1 - 6, with 1 being most interested)",
       title = "Aquaculture Topics Ranked by Interest in Learning More")


### need to add std dev or std error bars???
```

```{r}
# Q30_1 - Q30_6 (Rank the forms of communication by which you would prefer to learn about aquaculture, with 1 being most interested)

# q30_1 - short educational film
# q30_2 - physical pamphlets
# q30_3 - festivals/conferences
# q30_4 - website
# q30_5 - social media/blogs
# q30_6 - other
# q30_6_TEXT - (other text)

# make new subset of data
comm_data_2 <- survey_data_tidy %>% 
  select(q30_1:q30_6) %>% 
  mutate_if(is.character,as.numeric) %>% 
  summarize_all(mean, na.rm = TRUE) # calculate mean rank score for each answer choice

# transpose dataframe
comm_data_2 <- t(comm_data_2) %>% 
  as.data.frame() %>% 
  rownames_to_column(var = "form_of_communication") %>% 
  mutate(form_of_communication = case_when(
    form_of_communication == "q30_1" ~ "short educational films",
    form_of_communication == "q30_2" ~ "physical pamphlets",
    form_of_communication == "q30_3" ~ "festivals/conferences",
    form_of_communication == "q30_4" ~ "website",
    form_of_communication == "q30_5" ~ "social media/blogs",
    form_of_communication == "q30_6" ~ "other")) %>% 
  mutate(form_of_communication = fct_reorder(form_of_communication, V1, .desc = TRUE))  # order topics by count, smallest (most interested) to largest (least interested)

ggplot(data = comm_data_2, aes(x = form_of_communication, y = V1)) +
  geom_col(fill = "skyblue4") +
  coord_flip() +
  theme_minimal() +
  labs(x = "",
       y = "Average Ranked Score (from 1 - 6, with 1 being most interested)",
       title = "Ranked Forms of Communication")


### need to add std dev or std error bars???

```

```{r}
## Word Cloud list of "other" forms of communication

# convert each string of words to lowercase
survey_data_tidy$q30_6_text <- tolower(survey_data_tidy$q30_6_text)

other_comm_forms <- survey_data_tidy %>% 
  select(q30_6_text) %>%
  mutate(q30_6_text = str_trim(q30_6_text)) %>%  # remove excess whitespace before/after word
  count(q30_6_text) %>%                          # count repeating words
  arrange(-n)                                    # arrange word counts from largest to smallest

wordcloud_other_comm_forms <- ggplot(data = other_comm_forms, 
                                    aes(label = q30_6_text)) +
  geom_text_wordcloud(aes(color = n, size = n), 
                      shape = "diamond") +
  scale_size_area(max_size = 80) +
  scale_color_gradientn(colors = c("burlywood4", "green")) +
  theme_minimal() +
  labs(y = "other forms of communication") +
  theme(axis.line.y = element_line(color="black", size = 0.1))

wordcloud_other_comm_forms
```


## Citations: 

California county boundaries layer: [US Census Bureau's 2016 MAF/TIGER database.](https://data.ca.gov/dataset/ca-geographic-boundaries)

California ZIP code tabulations areas layer: [US Census Bureau's 2020 TIGER/Line Shapefiles database.](https://www.census.gov/cgi-bin/geo/shapefiles/index.php?year=2021&layergroup=ZIP+Code+Tabulation+Areas)

zipcode R package []
(https://cran.r-project.org/web/packages/zipcodeR/zipcodeR.pdf)
