==> Rmarkdowns/4_survey_analytical_tests_sr.Rmd <==
---
title: "Survey - Analytical Tests"
author: "Sydney Rilum"
date: "1/26/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(here)
library(janitor)
library(broom)
library(kableExtra)
library(data.table)  # advanced pivot_table tools
library(naniar)      # replace NA tools
library(zipcodeR)    # for zipcode assignment by county
library(RColorBrewer)
library(lattice)     # for heatmaps
#library(MASS)        # for polr function (used in ordered logistic regression)
#library(raster)      # for mapping raster layers
#library(dplyr)       # needed ONLY if working on a windows computer
```

## Data Set-up & Cleaning 

```{r}
# Read in combined survey data
survey_data <- read.csv(here("data","survey_data","combined_survey_responses.csv")) 

survey_data_clean <- survey_data %>%
  clean_names() %>% 
  slice(-1) # remove first row of df
```

```{r}
## Filter for Santa Barbara & Ventura County Zipcodes 
# Q37 - Zipcode Data

# Create a dataframe containing only zipcodes from Ventura and Santa Barbara Counties
## Not run:
download_zip_data() # uses zipcodeR data 
## End(Not run)

ventura_co_zip_full <-
  search_county("ventura", "CA") 

ventura_zipcodes<- 
  ventura_co_zip_full[,c("zipcode", "county", "lat", "lng")]

sb_co_zip_full <- 
 search_county("Santa Barbara", "CA")

sb_zipcodes<- 
  sb_co_zip_full[,c("zipcode", "county", "lat", "lng")]

# combine SB & Ventura zipcode data frames
sb_ventura_co_zip <- 
  rbind(ventura_zipcodes, sb_zipcodes)
```

```{r}
# select survey responses from Ventura County residents ONLY
survey_data_v <- merge(survey_data_clean, ventura_zipcodes, 
                         by.x = "q37", by.y = "zipcode")

# select survey responses from Santa Barbara County residents ONLY
survey_data_sb <- merge(survey_data_clean, sb_zipcodes, 
                         by.x = "q37", by.y = "zipcode")

# select survey responses from Ventura AND Santa Barbara County residents
survey_data_sbv <- merge(survey_data_clean, sb_ventura_co_zip, 
                         by.x = "q37", by.y = "zipcode")
```


```{r}
## Convert dataframe to tidydata format (Reshaping the data)

# convert data frame into a data table for ventura county zipcodes
survey_data_v <- as.data.table(survey_data_v)

# reshape data from wide to long format
survey_data_v_tidy <- survey_data_v %>% 
  melt(id.vars = c("response_id", "response_type", "finished"), 
       variable.name = "question", 
       value.name = "answer")


# repeat for sb county zipcodes
survey_data_sb <- as.data.table(survey_data_sb)

# reshape data from wide to long format
survey_data_sb_tidy <- survey_data_sb %>% 
  melt(id.vars = c("response_id", "response_type", "finished"), 
       variable.name = "question", 
       value.name = "answer")


# repeat for combined ventura AND sb county zipcodes
survey_data_sbv <- as.data.table(survey_data_sbv)

# reshape data from wide to long format
survey_data_sbv_tidy <- survey_data_sbv %>% 
  melt(id.vars = c("response_id", "response_type", "finished"), 
       variable.name = "question", 
       value.name = "answer")

```



```{r}
## Filter Survey Response Type

# Response Types Key:
# 0 = speeders (time removal)
# 1 = completes (approved by our team)
# 2 = removed by our team for replacement (flagged for not answering ranking questions, but otherwise fine)
# 3 = response collected via our own outreach (QRcode/email/social media)

# (Note: Qualtrics Paid Responses = 0,1,2)

# remove speeders responses (response type: 0) from survey dataframes - for ventura county zipcodes
survey_data_v_tidy_filtered <- survey_data_v_tidy %>% 
  filter(response_type != 0)

# repeat for sb county zipcodes
survey_data_sb_tidy_filtered <- survey_data_sb_tidy %>% 
  filter(response_type != 0)

# repeat for ventura AND sb county combined zipcodes
survey_data_sbv_tidy_filtered <- survey_data_sbv_tidy %>% 
  filter(response_type != 0)
```


# Chi-Square Test Assumptions & Hypothesis Testing

A chi-square test for independence will be used to compare proportions/count data collected in the survey and determine whether or not there is an association between word choice/terminology and a residents’ support for kelp aquaculture.

*Assumptions for Chi-Square Test:*

A Chi-Square test of independence is used to determine whether or not there is a significant association between two categorical variables.
This test makes four assumptions:

*Assumption 1:* Both variables are categorical.

It’s assumed that both variables are categorical. That is, both variables take on values that are names or labels.

Our variables: word choice/terminology (either kelp farming, seaweed farming, kelp aquaculture, or seaweed aquaculture) and support for kelp aquaculture (five likert scale categories from strongly opposed to strongly supportive)

*Assumption 2:* All observations are independent.

It’s assumed that every observation in the dataset is independent. That is, the value of one observation in the dataset does not affect the value of any other observation.

*Assumption 3:* Cells in the contingency table are mutually exclusive.

It’s assumed that individuals can only belong to one cell in the contingency table. That is, cells in the table are mutually exclusive – an individual cannot belong to more than one cell.

*Assumption 4:* Expected value of cells should be 5 or greater in at least 80% of cells.

It’s assumed that the expected value of cells in the contingency table should be 5 or greater in at least 80% of cells and that no cell should have an expected value less than 1.



# 1. & 2. Wording Experiment: 

*Question:* Does word choice/terminology influence residents' support for kelp aquaculture?

*Hypothesis:*  Yes, people will be more supportive of ‘kelp farming’ than of 'kelp aquaculture', 'seaweed aquaculture', and 'seaweed farming' because of its associations with an important native Californian species and since farming is a more familiar term.

*Statistical Test:* Chi-square test of Q2, Q4, Q6, Q8


## Chi-Square Test for Ventura County residents ONLY
```{r}
# Create a table with both counts & proportions
support_counts_v <- survey_data_v_tidy_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>% 
  mutate(number = case_when( 
    answer == "Strongly opposed" ~ 1,
    answer == "Somewhat opposed" ~ 2,
    answer == "Neither opposed nor in support" ~ 3,
    answer == "Somewhat supportive" ~ 4,
    answer == "Strongly supportive" ~ 5)) %>%
  mutate_all(na_if,"") %>% 
  drop_na()

support_counts_v_table <- support_counts_v %>% 
  tabyl(question, answer) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") 
```

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `question` column entries into stored rownames)
support_v_ct <- support_counts_v_table %>% 
  column_to_rownames(var = "question")

# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
support_v_x2 <- chisq.test(support_v_ct)
support_v_x2

# Get the chi-square test results to call outputs in-line
support_v_tidy <- tidy(support_v_x2)
```

The p-value of the test is 0.3301. Since this p-value is not less than .05, we do not have sufficient evidence to say that there is an association between word choice/terminology and support for kelp aquaculture, for Ventura County residents.

## Fisher exact test for Ventura County residents ONLY

The data violate the assumptions of the chi-square test (80% of the expected values >5). The fisher exact test will be used instead. 

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `q10` column entries into stored rownames)

fisher_support_v <- fisher.test(support_v_ct)
fisher_support_v <- tidy(fisher_support_v)
```

## Chi-Square Test for SB County residents ONLY
```{r}
# Create a table with both counts & proportions
support_counts_sb <- survey_data_sb_tidy_filtered %>% 
  select(question, answer) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ 1,
    answer == "Somewhat opposed" ~ 2,
    answer == "Neither opposed nor in support" ~ 3,
    answer == "Somewhat supportive" ~ 4,
    answer == "Strongly supportive" ~ 5)) %>%
  mutate_all(na_if,"") %>% 
  drop_na()

support_counts_sb_table <- support_counts_sb %>% 
  tabyl(question, answer) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") 
```

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `question` column entries into stored rownames)
support_sb_ct <- support_counts_sb_table %>% 
  column_to_rownames(var = "question")

# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
support_sb_x2 <- chisq.test(support_sb_ct)
support_sb_x2

# Get the chi-square test results to call outputs in-line
support_sb_tidy <- tidy(support_sb_x2)
```

The p-value of the test is 0.4502. Since this p-value is not less than .05, we do not have sufficient evidence to say that there is an association between word choice/terminology and support for kelp aquaculture, for Santa Barbara County residents.

## Fisher exact test for SB County residents ONLY

The data violate the assumptions of the chi-square test (80% of the expected values >5). The fisher exact test will be used instead. 

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `q10` column entries into stored rownames)

fisher_support_sb <- fisher.test(support_sb_ct)
fisher_support_sb <- tidy(fisher_support_sb)
```

Since the chi-square tests revealed no association between word choice/terminology and support for kelp aquaculture for both Ventura and Santa Barbara County residents individually, we can now combine the datasets and perform a chi-square test once more for further analysis.




















#### USE THIS CODE BELOW!!! 

```{r}
# load in data (CSV created in survey_data_weighting.Rmd)
survey_data_tidy_sbv_filtered <- read.csv(here("data","survey_data_tidy_sbv_filtered.csv"))
survey_data_tidy_sbv_long_filtered <- read.csv(here("data","survey_data_tidy_sbv_long_filtered.csv"))
```


# 1. & 2. Wording Experiment: 

*Question:* Does word choice/terminology influence residents' support for kelp aquaculture?

*Hypothesis:*  Yes, people will be more supportive of ‘kelp farming’ than of 'kelp aquaculture', 'seaweed aquaculture', and 'seaweed farming' because of its associations with an important native Californian species and since farming is a more familiar term.

*Statistical Test:* Chi-square test of Q2, Q4, Q6, Q8


## Chi-Square Test for SB & Ventura County residents combined
```{r}
# Create a table with both counts & proportions
support_counts <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>% 
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight))

support_counts_table <- support_counts %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0))

```

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `question` column entries into stored rownames)
support_ct <- support_counts_table %>% 
  column_to_rownames(var = "question")

# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
support_x2 <- chisq.test(support_ct)
support_x2

# Get the chi-square test results to call outputs in-line
support_tidy <- tidy(support_x2)
```

The p-value of the test is 0.3543. Since this p-value is less than .05, we have sufficient evidence to say that there is an association between word choice/terminology and support for kelp aquaculture.

There is a significant association between word choice and whether a Santa Barbara and Ventura County residents are supportive of kelp aquaculture ($\chi$^2^(`r support_tidy$parameter`) = `r round(support_tidy$statistic,2)`, *p* = `r format(support_tidy$p.value, scientific = TRUE, digits = 3)`). Therefore, we reject the null hypothesis that there is no association between word choice and whether a Santa Barbara or Ventura County resident is supportive of kelp aquaculture. 



## Chi-Square Test for SB & Ventura County residents combined (WITH COMBINING ANSWER CHOICES)
```{r}
# Create a table with both counts & proportions
support_counts_combined <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ "Opposed",
    answer == "Somewhat opposed" ~ "Opposed",
    answer == "Neither opposed nor in support" ~ "Neither opposed nor in support",
    answer == "Somewhat supportive" ~ "Supportive",
    answer == "Strongly supportive" ~ "Supportive")) %>%
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight))

support_counts_combined_table <- support_counts_combined %>%  
  pivot_wider(names_from = answer,
              values_from = weight_count) 

```

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `question` column entries into stored rownames)
support_combined_ct <- support_counts_combined_table %>% 
  column_to_rownames(var = "question")

# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
support_combined_x2 <- chisq.test(support_combined_ct)
support_combined_x2

# Get the chi-square test results to call outputs in-line
support_combined_tidy <- tidy(support_combined_x2)
```

The p-value of the test is 0.03636. Since this p-value is less than .05, we have sufficient evidence to say that there is a significant association between word choice/terminology and support for kelp aquaculture.

X2 (6, N = 350) = 13.455, p > .03636

There is  a significant association between word choice and whether a Santa Barbara and Ventura County residents are supportive of kelp aquaculture ($\chi$^2^(`r support_tidy$parameter`) = `r round(support_tidy$statistic,2)`, *p* = `r format(support_tidy$p.value, scientific = TRUE, digits = 3)`). Therefore, we reject the null hypothesis that there is no association between word choice and whether a Santa Barbara or Ventura County resident is supportive of kelp aquaculture. 



AND thus we can now group answers from Q2&4 and Q6&8 together for further analyses, assuming that this combined data set reflects respondents support or opposition to the synonymous terms of kelp/seaweed aquaculture and kelp/seaweed farming.


## Fisher exact test 
## not needed unless violate Chi-square results 

The data violate the assumptions of the chi-square test (80% of the expected values >5). The fisher exact test will be used instead. 

```{r}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `q10` column entries into stored rownames)

#fisher_support_sbv <- fisher.test(support_ct) # , workspace = 2e8)
#fisher_support_sbv <- tidy(fisher_support_sbv)
```





# 3. Perception: 

*Question:* Does perception of aquaculture change when specifying seaweed aquaculture? 

*Hypothesis:* Yes, perception of seaweed aquaculture will be more positive than aquaculture more broadly. 

*Statistical Test:* Chi-square test of Q12 vs. Q13

```{r, perception}
# Pull Q12 and Q13 on perception of aquaculture and seaweed aquaculture, assign numerical value to categorical response
perception_counts <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight) %>% 
  filter(question == "q12" | 
         question == "q13") %>% 
  mutate(number = case_when( 
    answer == "Very Negative" ~ 1,
    answer == "Negative" ~ 2,
    answer == "Neutral" ~ 3,
    answer == "Positive" ~ 4,
    answer == "Very Positive" ~ 5,
    answer == "Unsure/Not familiar" ~ 6)) %>%
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(question, answer) %>% 
  summarize(weight_count = sum(weight))

perception_counts_table <- perception_counts %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) 
```


```{r, perception chi-square test}
# Use `column_to_rownames` to convert counts table to a contingency table (makes `question` column entries into stored rownames)
perception_ct <- perception_counts_table %>% 
  column_to_rownames(var = "question")

# Chi-square test for independence: Is there an association between a more positive perception of aquaculture and the specification of aquaculture being for seaweed?
perception_x2 <- chisq.test(perception_ct)
perception_x2

# Get the chi-square test results to call outputs in-line
perception_tidy <- tidy(perception_x2)
```

The p-value of the test is 0.4925. Since this p-value is not less than .05, we do not have sufficient evidence to say that there is a significant association between word choice/terminology and perception.

X2 (5, n = 350) = 4.4064, p = 0.4925


```{r, perception t-test NOT USING}
# get means
#perception_q12<- perception_counts_table %>% 
#  filter(question == "q12") %>% 
#  column_to_rownames("question") %>% 
#  toString() #%>% 
#  #pull(weight_count)
#perception_q13 <- perception_counts_table %>% 
#  filter(question == "q13") %>% 
#  column_to_rownames("question") %>% 
#  toString()
#  #pull(weight_count)

# run t-test
#perception_ttest <- t.test(perception_q12, perception_q13)
#perception_ttest

#perception_ttest_tidy <-tidy(perception_ttest)

#mean(perception_q12)
#mean(perception_q13)
```

The p-value of the Welch Two-sample t-test is 0.61. Since this p-value is not less than .05, we do not have sufficient evidence to say that there is a statistically significant difference between the perception of aquaculture and seaweed aquaculture.

There is not a significant difference in perception whether a Santa Barbara and Ventura County resident is supportive of aquaculture when seaweed is specified as the type, *p* = `r format(perception_ttest_tidy$p.value, scientific = TRUE, digits = 3)`). Therefore, we fail to reject the alternative hypothesis that seaweed would make respondents perception more positive and we accept the null hypothesis that there is no significant difference between aquaculture being specified as seaweed aquaculture. 






# 4. Relationship Between Perception and Support: 

*Question:* Is support related to more positive perceptions of seaweed aquaculture?

*Hypothesis:* Yes, supporters will have more positive perceptions of seaweed aquaculture.

*Assumptions:* 
- Since Q2,4, 6, 8 are not significantly different, then questions/answers can thus be combined.
- Remove "don't know" answer choice counts

*Statistical Test:* Ordinal logistic regression of perception combined support values (Q2,4,6,8) vs. perception value from Q13 (seaweed aquaculture)


dependent variable = support (Q2)
independent variable = perception (Q13)


```{r}
# create data frame with Q2 and Q13

ordinal_log_reg_data <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8" |
         question == "q13") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2",
    question == "q13" ~ "q13")) %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ -2,
    answer == "Somewhat opposed" ~ -1,
    answer == "Neither opposed nor in support" ~ 0,
    answer == "Somewhat supportive" ~ 1,
    answer == "Strongly supportive" ~ 2,
    answer == "Very Negative" ~ -2,
    answer == "Negative" ~ -1,
    answer == "Neutral" ~ 0,
    answer == "Positive" ~ 1,
    answer == "Very Positive" ~ 2,
    answer == "Unsure/Not familiar" ~ -999)) %>%
  replace_with_na(replace = list(answer = -999))%>%  # remove "unsure" count data
  filter(answer != "NA") # remove incomplete responses

```

```{r}
# use spread() to pivot table to wide format
ordinal_log_reg <- ordinal_log_reg_data %>% 
  select(response_id, question, answer) %>% 
  spread(question, answer) %>% 
  filter(q13 != "NA")  %>% # remove incomplete responses
  filter(q2 != "NA") # remove incomplete responses
  
ordinal_log_reg_table<- ordinal_log_reg %>% 
  tabyl(q2, q13) %>% 
  column_to_rownames(var = "q2")
```



# 5. Knowledge/Familiarity: 

*Question:* Does familiarity with aquaculture influence their perception of aquaculture? 

*Null Hypothesis (H0):*  There is no association between the two variables (familiarity and perceptions) i.e. they are independent 

*Statistical Test:* Chi-square test of Q10 & Q13

*Test Statistic:* P-value > 0.05 


#### Q10 (familiarity) v. Q13 (perception of seaweed aq)

```{r 10v.13 dataframe, table and plot WEIGHTED}
#create subset from tidy filtered data
#keep only question 10 and 13
k_p_weighted <- survey_data_tidy_sbv_long_filtered %>% 
  select(response_id, question, answer, weight) %>% 
  filter(question == "q10"|
           question == "q13") %>% 
  #switch to wide format with response id as the link
  spread(question, answer) %>%
  #mutate(q13 = case_when(
  #  q13 == "Very Negative" ~ "Negative",
  #  q13 == "Very Positive" ~ "Positive",
  #  TRUE ~ q13)) %>% 
  mutate(q13 = fct_relevel(q13, c(c("Very Negative", "Negative","Neutral","Positive", "Very Positive", "Unsure/Not familiar")))) %>% 
  #mutate(q13 = fct_relevel(q13, c(c("Negative","Neutral","Positive", "Unsure/Not familiar")))) %>% 
  mutate(q10 = fct_relevel(q10, c(c("Very familiar", "Basic understanding", "Heard of it, but don't know details", "Never heard of it")))) %>%  
  group_by(q13, q10) %>% 
  summarize(weight_count = round(sum(weight),)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

## graph
ggplot(k_p_weighted, aes(fill= q13, x = q10, y = weight_count))+
  geom_col(position = position_dodge2(preserve = "single")) +
  scale_x_discrete(labels= function(x) str_wrap(x, width = 10)) +
  scale_y_continuous(expand = expansion(mult = c(0.005, 0))) +
  scale_fill_manual(values =c("#D7191C", "#FDAE61", "#FFFFBF", "#ABD9E9","#2C7BB6", "#999999")) +
  #scale_fill_manual(values =c("#FDAE61", "#FFFFBF", "#ABD9E9", "#999999")) +
  labs(y = "Response Count",
       x = "Familiarity with Aquaculture",
       title = "Relationship Between Familiarity and Perception",
       fill = "Perception of \nSeaweed Aquaculture") + 
  theme_minimal() +
  theme(axis.line.y = element_line(color = "black", size = 0.5),
        axis.text.y = element_text(color = "black"),
        axis.line.x = element_line(color = "black", size = 0.5),
        axis.text.x = element_text(color = "black")) 


ggsave(filename = file.path(here("graphs","familiarity_&_perception.png")))
```

One key assumption of the chi-square test is that 80% of the cells must have an expected value of 5 or greater. Only 14 of the 24 cells meet this requirement (58%), therefore violating the assumptions of the chi-square test. We will use the maximum liklihood chi-square test to also understand the relationship between the two variables. 

```{r 10 v. 13 chi squared test, violates assumptions!!!}
#put counts into table format
k_p_ct_table <- k_p_weighted %>% 
  select(-proportion) %>% 
  pivot_wider(names_from = q10, 
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "q13")

# Chi-square test for independence: Is there an association between familiarity and perception?
#knowledge_perception_weight_x2 <- chisq.test(k_p_ct_table)
#knowledge_perception_weight_x2

# Get the chi-square test results to call outputs in-line
#k_p_x2_tidy <- tidy(knowledge_perception_weight_x2)

fisher.test(k_p_ct_table, simulate.p.value=TRUE)
```


The p-squared value `r round(k_p_x2_tidy[1, 2], 2)` is less than alpha (0.05) indicating the null hypothesis- that the two variables are independent- failed. However, there were some counts that were less than 5 which indicates there should be additional study. Thus, the data violates an assumption of the chi-square test so we conduct the maximum likelihood test below. 

We will aggregate 'very negative' and 'negative' and 'very positive' and 'postiive' and run the chi-square with these groups. 

```{r group Q13 responses, to meet chi-square assumption}
#create subset from tidy filtered data
#keep only question 10 and 13
k_p_grouped_weighted <- survey_data_tidy_sbv_long_filtered %>% 
  select(response_id, question, answer, weight) %>% 
  filter(question == "q10"|
           question == "q13") %>% 
 #switch to wide format with response id as the link
  spread(question, answer) %>% 
 #group responses
  mutate(q13= case_when(
    q13 == "Very Negative" ~ "Negative",
    q13 == "Negative" ~ "Negative",
    q13 =="Neutral" ~ "Neutral",
    q13 == "Positive" ~ "Positive",
    q13 =="Very Positive" ~ "Positive",
    q13 =="Unsure/Not familiar" ~ "Unsure/Not familiar")) %>% 
  mutate(q13 = fct_relevel(q13, c(c("Negative","Neutral","Positive", "Unsure/Not familiar")))) %>% 
  mutate(q10 = fct_relevel(q10, c(c("Very familiar", "Basic understanding", "Heard of it, but don't know details", "Never heard of it")))) %>%  
  group_by(q13, q10) %>% 
  summarize(weight_count = sum(weight)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

k_p_ct_table_grouped <- k_p_grouped_weighted %>% 
  select(-proportion) %>% 
  pivot_wider(names_from = q10, 
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "q13")

# Chi-square test for independence: Is there an association between familiarity and perception?
knowledge_perception_grouped_x2 <- chisq.test(k_p_ct_table_grouped)
knowledge_perception_grouped_x2


# Get the chi-square test results to call outputs in-line
k_p_x2_tidy <- tidy(knowledge_perception_grouped_x2)
```


When perception is combined into 'negative', 'neutral', 'positive', and 'unsure', the null hypothesis is rejected (p-value= 2.22 x10-7, alpha =0.05) which indicates the variables are dependent. 

We can also use the Fisher's exact test to calculate the exact p-value. **Fisher's test requires a high number of simulations because the value of p is so small. Kept crashing R**

```{r 10 v. 13 Fischer's exact test}
#fisher.test(k_p_ct, workspace = 2e7)
```











#6. Relationship Between Knowledge and Support: 

*Question:* Does familiarity with aquaculture influence support for expansion? 

*Hypothesis:* Those who are familiar with aquaculture will have more polarized opinions. Those less familiar will have more positive views of kelp aquaculture. 

*Statistical Test:* Ordinal logistic regression of familiarity (Q10, Q16, and/or Q17) to predict support (combined result of support: Q2,4,6,8)

```{r}
# create data frame with Q2 and Q10

ordinal_log_reg_2_data <- survey_data_tidy_sbv_long_filtered %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8" |
         question == "q10") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2",
    question == "q10" ~ "q10")) %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ -2,
    answer == "Somewhat opposed" ~ -1,
    answer == "Neither opposed nor in support" ~ 0,
    answer == "Somewhat supportive" ~ 1,
    answer == "Strongly supportive" ~ 2,
    answer == "Very familiar" ~ 1,
    answer == "Basic understanding" ~ 2,
    answer == "Heard of it, but don't know details" ~ 3,
    answer == "Never heard of it" ~ 4)) %>%
  filter(answer != "NA") # remove incomplete responses

```

```{r}
# use spread() to pivot table to wide format
ordinal_log_reg_2 <- ordinal_log_reg_2_data %>% 
  select(response_id, question, answer) %>% 
  spread(question, answer) %>% 
  filter(q10 != "NA")  %>% # remove incomplete responses
  filter(q2 != "NA") # remove incomplete responses
  
ordinal_log_reg_2_table<- ordinal_log_reg_2 %>% 
  tabyl(q2, q10) %>% 
  column_to_rownames(var = "q2")
```








## Regression Models 

####Relationship Between perception and support

*Question:* Is support related to more postie perceptions of seaweed aquaculture?

*Hypothesis:* Those who view seaweed aquaculture more positively will be more likely to be in support of seaweed aquaculture expansion.

*Assumptions:* 
- Since Q2,4, 6, 8 are not significantly different, then questions/answers can thus be combined.
- Remove "don't know" answer choice counts

*Statistical Test:* Ordered logistic regression (OLR) of perception combined support values (Q2,4,6,8) vs. perception of seaweed aquaculture (Q13)

dependent variable = support (Q2)
independent variable = perception (Q13)


```{r regression table }
# create data frame with Q2 and Q13 

OLR_perception_support_data <- survey_data_tidy_sbv_long_filtered %>% 
  select(response_id, question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8" |
         question == "q13") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2",
    question == "q13" ~ "q13")) %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ -2,
    answer == "Somewhat opposed" ~ -1,
    answer == "Neither opposed nor in support" ~ 0,
    answer == "Somewhat supportive" ~ 1,
    answer == "Strongly supportive" ~ 2,
    answer == "Very Positive" ~ 2,
    answer == "Positive" ~ 1,
    answer == "Neutral" ~ 0,
    answer == "Negative" ~ -1,
    answer == "Very Negative" ~ -2,
    answer == "Unsure/Not familiar" ~ -999)) %>%
  replace_with_na(replace = list(answer = -999))%>% 
  filter(answer != "NA") # remove incomplete responses

```

```{r}
# use spread() to pivot table to wide format
OLR_perception_support <- OLR_perception_support_data %>% 
  spread(question, answer) %>% 
  filter(q13 != "NA")  %>% # remove incomplete responses
  filter(q2 != "NA") # remove incomplete responses
  
OLR_perception_support_table <- OLR_perception_support %>% 
  tabyl(q2, q13) %>% 
  column_to_rownames(var = "q2")

summary(OLR_perception_support_table)

```

## repeat but with grouped support for aquaculture and farming

```{r regression table}
# create data frame with Q2 x2  and Q13 

OLR_perception_support_data <- survey_data_tidy_sbv_long_filtered %>% 
  select(response_id, question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8" |
         question == "q13") %>%
  mutate(question = case_when( 
    question == "q2" ~ "aquaculture",
    question == "q4" ~ "aquaculture",
    question == "q6" ~ "farming",
    question == "q8" ~ "farming",
    question == "q13" ~ "perception")) %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ -2,
    answer == "Somewhat opposed" ~ -1,
    answer == "Neither opposed nor in support" ~ 0,
    answer == "Somewhat supportive" ~ 1,
    answer == "Strongly supportive" ~ 2,
    answer == "Very Positive" ~ 2,
    answer == "Positive" ~ 1,
    answer == "Neutral" ~ 0,
    answer == "Negative" ~ -1,
    answer == "Very Negative" ~ -2,
    answer == "Unsure/Not familiar" ~ -999)) %>%
  replace_with_na(replace = list(answer = -999))%>% 
  filter(answer != "NA") # remove incomplete responses

```

```{r}
# aquaculture vs. perception
# make dataframe one for each regression
perception_support_aquaculture <- OLR_perception_support_data %>% 
  spread(question, answer) %>% 
  select(aquaculture, perception, weight) %>% 
  filter(perception != "NA",
         aquaculture != "NA") # remove incomplete responses

# make contingency table
perception_support_aquaculture_table <- perception_support_aquaculture %>% 
  group_by(aquaculture, perception) %>% 
  summarize(weight_count = round(sum(weight),0)) %>% 
  pivot_wider(names_from = perception,
              values_from = weight_count) %>% 
  column_to_rownames(var = "aquaculture") %>% 
  mutate_all(~replace(., is.na(.), 0))

summary(perception_support_aquaculture_table)

# repeat for farming vs. perception
perception_support_farming <- OLR_perception_support_data %>% 
  spread(question, answer) %>% 
  select(farming, perception, weight) %>% 
  filter(perception != "NA",
         farming != "NA") # remove incomplete responses

perception_support_farming_table <- perception_support_farming %>% 
  group_by(farming, perception) %>% 
  summarize(weight_count = round(sum(weight),0)) %>% 
  pivot_wider(names_from = perception,
              values_from = weight_count) %>% 
  column_to_rownames(var = "farming") %>% 
  mutate_all(~replace(., is.na(.), 0))

summary(perception_support_farming_table)
```

```{r create ordinal logistic model}
library(MASS)
#use polr to create a logit model to understand how familiarity influences expansion
m1 <- polr(factor(aquaculture) ~ perception, data = perception_support_aquaculture, Hess =TRUE)
summary(m1)

m2 <- polr(factor(farming) ~ perception, data = perception_support_farming, Hess =TRUE)
summary(m2) 
```

the q13 coefficient value represents the eta so as eta gets bigger then zeta gets smaller. 

Intercepts are the zeta values. 

```{r understand odds ratio}
#interpret odds ratio to probabilities. Outside parenthesis are important 
(ci1 <- confint(m1))
exp(cbind(coef(m1), t(ci1)))

(ci2 <- confint(m2))
exp(cbind(coef(m2), t(ci2)))
```


```{r predicted probabilities to understand odds ratio}
newdf_1 <- data.frame(perception =c(-2,-1,0,1,2))
(phat1 <- predict(object = m1, newdf_1, type = "p"))

newdf_2 <- data.frame(perception =c(-2,-1,0,1,2))
(phat2 <- predict(object = m2, newdf_2, type = "p"))

coul <- colorRampPalette(brewer.pal(8, "Blues"))(25)

# heatmap plots
levelplot(phat1, col.regions= coul,
          xlab = "Perception",
          ylab = "Level of Support",
          main = "Aquaculture",
          at = seq(min(0), max(1), length.out=20)) # make scale 0 to 1

levelplot(phat2, col.regions= coul,
          xlab = "Perception",
          ylab = "Level of Support",
          main = "Farming",
          at = seq(min(0), max(1), length.out=20))

phat1 %>% 
  t() %>% 
  kable(#align = c("c", "c", "c", "c", "c"),
        row.names = TRUE,
        digits = 2,
        caption = "Predicted Probabilities of Support for Aquaculture as a Function of Perception") %>%
  #row_spec(0, align = "c") %>%
  kable_classic(full_width = T) %>%
  kable_styling(bootstrap_options = "striped")

phat2 %>% 
  t() %>% 
  kable(#align = c("c", "c", "c", "c", "c"),
        row.names = TRUE,
        digits = 2,
        caption = "Predicted Probabilities of Support for Farming as a Function of Perception") %>%
  #row_spec(0, align = "c") %>%
  kable_classic(full_width = T) %>%
  kable_styling(bootstrap_options = "striped")
```

```{r}
# unload MASS package, to remove conflicts with "dplyr::select" function
detach("package:MASS", unload=TRUE)
```


####Relationship Between Familiarity and Support: 


*Question:* Does familiarity with seaweed aquaculture depend on whether or not one is in support of aquaculture?

*Hypothesis:* Yes, supporters will be more familiar with seaweed aquaculture.

*Assumptions:* 
- Since Q2,4, 6, 8 are not significantly different, then questions/answers can thus be combined.
- Remove "don't know" answer choice counts

*Statistical Test:* Ordinal logistic regression of perception combined support values (Q2,4,6,8) vs. perception value from Q13 (seaweed aquaculture)


dependent variable = support (Q2)
independent variable = familiarity (Q10)


```{r regression table }
# create data frame with Q2 x 2 and Q10

OLR_familiarity_support_data <- survey_data_tidy_sbv_long_filtered %>% 
  select(response_id, question, answer, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8" |
         question == "q10") %>%
  mutate(question = case_when( 
    question == "q2" ~ "aquaculture",
    question == "q4" ~ "aquaculture",
    question == "q6" ~ "farming",
    question == "q8" ~ "farming",
    question == "q10" ~ "familiarity")) %>% 
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~ -2,
    answer == "Somewhat opposed" ~ -1,
    answer == "Neither opposed nor in support" ~ 0,
    answer == "Somewhat supportive" ~ 1,
    answer == "Strongly supportive" ~ 2,
    answer == "Very familiar" ~ 4,
    answer == "Basic understanding" ~ 3,
    answer == "Heard of it, but don't know details" ~ 2,
    answer == "Never heard of it" ~ 1)) %>%
  filter(answer != "NA") # remove incomplete responses

```

```{r}
# aquaculture vs. familiarity
# make dataframe one for each regression
familiarity_support_aquaculture <- OLR_familiarity_support_data %>% 
  spread(question, answer) %>%  # use spread() to pivot table to wide format
  select(aquaculture, familiarity, weight) %>% 
  filter(familiarity != "NA",
         aquaculture != "NA") # remove incomplete responses

# make contingency table
familiarity_support_aquaculture_table <- familiarity_support_aquaculture %>% 
  group_by(aquaculture, familiarity) %>% 
  summarize(weight_count = round(sum(weight),0)) %>% 
  pivot_wider(names_from = familiarity,
              values_from = weight_count) %>% 
  column_to_rownames(var = "aquaculture") %>% 
  mutate_all(~replace(., is.na(.), 0))

summary(familiarity_support_aquaculture_table)

# repeat for farming vs. perception
familiarity_support_farming <- OLR_familiarity_support_data %>% 
  spread(question, answer) %>% 
  select(farming, familiarity, weight) %>% 
  filter(familiarity != "NA",
         farming != "NA") # remove incomplete responses

familiarity_support_farming_table <- familiarity_support_farming %>% 
  group_by(farming, familiarity) %>% 
  summarize(weight_count = round(sum(weight),0)) %>% 
  pivot_wider(names_from = familiarity,
              values_from = weight_count) %>% 
  column_to_rownames(var = "farming") %>% 
  mutate_all(~replace(., is.na(.), 0))

summary(familiarity_support_farming_table)
```

```{r create ordinal logistic model}
library(MASS)
#use polr to create a logit model to understand how familiarity influences expansion
m3 <- polr(factor(aquaculture) ~ familiarity, data = familiarity_support_aquaculture, Hess =TRUE)
summary(m3)

m4 <- polr(factor(farming) ~ familiarity, data = familiarity_support_farming, Hess =TRUE)
summary(m4)
```

```{r understand odds ratio}
#interpret odds ratio to probabilities. Outside parenthesis are important 
(ci3 <- confint(m3))
exp(cbind(coef(m3), t(ci3)))

(ci4 <- confint(m4))
exp(cbind(coef(m4), t(ci4)))
```

```{r predicted probabilities to understand odds ratio}
newdf_3 <- data.frame(familiarity =c(1,2,3,4))
(phat3 <- predict(object = m3, newdf_3, type = "p"))

newdf_4 <- data.frame(familiarity =c(1,2,3,4))
(phat4 <- predict(object = m4, newdf_4, type = "p"))

coul <- colorRampPalette(brewer.pal(8, "Blues"))(25)

# heatmap plots
levelplot(phat3, col.regions= coul,
          xlab = "Familiarity",
          ylab = "Level of Support",
          main = "Aquaculture",
          at = seq(min(0), max(1), length.out=20))

levelplot(phat4, col.regions= coul,
          xlab = "Familiarity",
          ylab = "Level of Support",
          main = "Farming",
          at = seq(min(0), max(1), length.out=20))

phat3 %>% 
  t() %>% 
  kable(#align = c("c", "c", "c", "c"),
        row.names = TRUE,
        digits = 2,
        caption = "Predicted Probabilities of Support for Aquaculture as a Function of Familiarity") %>%
  #row_spec(0, align = "c") %>%
  kable_classic(full_width = T) %>%
  kable_styling(bootstrap_options = "striped")

phat4 %>% 
  t() %>% 
  kable(#align = c("c", "c", "c", "c"),
        row.names = TRUE,
        digits = 2,
        caption = "Predicted Probabilities of Support for Farming as a Function of Familiarity") %>%
  #row_spec(0, align = "c") %>%
  kable_classic(full_width = T) %>%
  kable_styling(bootstrap_options = "striped")
```


```{r}
# unload MASS package, to remove conflicts with "dplyr::select" function
detach("package:MASS", unload=TRUE)
```


####Relationship Between Perception and Familiarity: 

*Question:* Is perception related to familiarity with seaweed aquaculture?

*Hypothesis:* 

*Assumptions:* 
- Remove "don't know" answer choice counts

*Statistical Test:* Ordinal logistic regression of 


dependent variable = perception (q13)
independent variable = familiarity (Q10)


```{r regression table }
# create data frame with Q10 and Q13

OLR_perception_familiarity_data <- survey_data_tidy_sbv_long_filtered %>% 
  select(response_id, question, answer, weight) %>% 
  filter(question == "q10" | 
         question == "q13") %>%
  mutate(question = case_when( 
    question == "q13" ~ "perception",
    question == "q10" ~ "familiarity")) %>% 
  mutate(answer = case_when( 
    answer == "Very familiar" ~ 4,
    answer == "Basic understanding" ~ 3,
    answer == "Heard of it, but don't know details" ~ 2,
    answer == "Never heard of it" ~ 1,
    answer == "Very Positive" ~ 2,
    answer == "Positive" ~ 1,
    answer == "Neutral" ~ 0,
    answer == "Negative" ~ -1,
    answer == "Very Negative" ~ -2,
    answer == "Unsure/Not familiar" ~ -999)) %>%
  replace_with_na(replace = list(answer = -999))%>% 
  filter(answer != "NA") # remove incomplete responses

```

```{r}
# make dataframe for regression perception vs. familiarity
OLR_perception_familiarity <- OLR_perception_familiarity_data %>%
  spread(question, answer) %>%  # use spread() to pivot table to wide format
  select(perception, familiarity, weight) %>% 
  filter(familiarity != "NA",
         perception != "NA") # remove incomplete responses
  
# make contingency table
OLR_perception_familiarity_table <- OLR_perception_familiarity %>% 
  group_by(perception, familiarity) %>% 
  summarize(weight_count = round(sum(weight),0)) %>% 
  pivot_wider(names_from = familiarity,
              values_from = weight_count) %>% 
  column_to_rownames(var = "perception") %>% 
  mutate_all(~replace(., is.na(.), 0))

summary(OLR_perception_familiarity_table)
```


```{r create ordinal logistic model}
library(MASS)
#use polr to create a logit model to understand how familiarity influences expansion
m5 <- polr(factor(perception) ~ familiarity, data = OLR_perception_familiarity, Hess =TRUE)
summary(m5)
```

```{r understand odds ratio}
#interpret odds ratio. Outside parenthesis are important 
(ci5 <- confint(m5))
exp(cbind(coef(m5), t(ci5)))
```

```{r predicted probabilities to understand odds ratio}
newdf5 <- data.frame(familiarity =c(1, 2, 3, 4))
(phat5 <- predict(object = m5, newdf5, type = "p"))

#coul is the color palette
coul <- colorRampPalette(brewer.pal(8, "Blues"))(100)

#heatmap
levelplot(phat5, col.regions= coul,
          xlab = "Familiarity",
          ylab = "Perception",
          main = "Predicted Probabilities",
          at = seq(min(0), max(1), length.out=20)) 

phat5 %>% 
  t() %>% 
  kable(#align = c("c", "c", "c", "c"),
        row.names = TRUE,
        digits = 2,
        caption = "Predicted Probabilities of Perception as a Function of Familiarity") %>%
  #row_spec(0, align = "c") %>%
  kable_classic(full_width = T) %>%
  kable_styling(bootstrap_options = "striped")
```


## Demogrphic Analyses - Association with Support for Seaweed Aquaculture

### chi-square test for association between COUNTY of residence and support for seaweed aquaculture

```{r}
# split up support/opposed responses by county
county_x2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight, county) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(county, answer) %>% 
  summarize(weight_count = round(sum(weight)),) 

county_x2_table <- county_x2 %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "county")
```

```{r}
# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
county_x2_result <- chisq.test(county_x2_table)
county_x2_result

# Get the chi-square test results to call outputs in-line
county_x2_results_tidy <- tidy(county_x2_result)

#fisher.test(county_x2_table)
```

### fisher's exact test for association between AGE of residence and support for seaweed aquaculture

data did not meet assumptions for chi-square test, even when support categories were combined.
therefore, a fisher's exact test was used instead.

```{r}
# split up support/opposed responses by county
age_x2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, age_group, weight) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(age_group, answer) %>% 
  summarize(weight_count = round(sum(weight)),) 

age_x2_table <- age_x2 %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "age_group")
```

```{r}
# fisher's exact test for independence: Is there an association between support for aquaculture and word choice?
fisher.test(age_x2_table, simulate.p.value=TRUE)
```


### fisher's exact for association between EDUCATION and support for seaweed aquaculture

data did not meet assumptions for chi-square test, even when support categories were combined.
therefore, a fisher's exact test was used instead.

```{r}
# split up support/opposed responses by county
education_x2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight, q34) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  filter(answer != "") %>% 
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(q34, answer) %>% 
  summarize(weight_count = round(sum(weight)),) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

education_x2_table <- education_x2 %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "q34")
```

```{r}
# Fisher's exact test Is there an association between support for aquaculture and word choice?
education_fisher_result <- fisher.test(education_x2_table, simulate.p.value=TRUE)
education_fisher_result
```




### chi-square test for association between POLIT AFFIL and support for seaweed aquaculture

```{r}
# split up support/opposed responses by county
polit_x2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight, q38) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  # combine categories to meet assumptions of test
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~"Opposed",
    answer == "Somewhat opposed" ~ "Opposed",
    answer == "Neither opposed nor in support" ~ "Neither opposed nor in support",
    answer == "Somewhat supportive" ~ "Supportive",
    answer == "Strongly supportive" ~ "Supportive")) %>%
  filter(answer != "") %>% 
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(q38, answer) %>% 
  summarize(weight_count = round(sum(weight)),) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

polit_x2_table <- polit_x2 %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "q38")
```

```{r}
# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
polit_result <- chisq.test(polit_x2_table)
polit_result

polit_results_tidy <- tidy(polit_result)

#fisher.test(polit_x2_table)
```


### chi-square test for association between SEX and support for seaweed aquaculture

```{r}
# split up support/opposed responses by county
sex_x2 <- survey_data_tidy_sbv_long_filtered %>% 
  select(question, answer, weight, q32) %>% 
  filter(question == "q2" | 
         question == "q4" |
         question == "q6" |
         question == "q8") %>%
  mutate(question = case_when( 
    question == "q2" ~ "q2",
    question == "q4" ~ "q2",
    question == "q6" ~ "q2",
    question == "q8" ~ "q2")) %>% 
  mutate(q32 = case_when( 
    q32 == "Man" ~ "Male",
    q32 == "Woman" ~ "Female")) %>% 
  filter(answer != "") %>% 
  mutate_all(na_if,"") %>% 
  drop_na() %>% 
  group_by(q32, answer) %>% 
  summarize(weight_count = round(sum(weight)),) 

sex_x2_table <- sex_x2 %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "q32")
```

```{r}
# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
sex_result <- chisq.test(sex_x2_table)
sex_result

sex_results_tidy <- tidy(sex_result)
```


### chi-square test for association between DAYS BY OCEAN and support for seaweed aquaculture

```{r}
# split up support/opposed responses by county
ocean_x2 <- survey_data_tidy_sbv_filtered %>% 
  select(response_id, q15, q2, q4, q6, q8, weight) %>% 
  #change non-responses to '0 days' because but one non-response corresponds to 'no' response to q14 'do you spend free time by the ocean?'
  mutate(q15= case_when(
    q15 == "" ~ "0 days", 
    TRUE ~ q15)) %>% 
  #combine support metrics from multiple columns into one columns
  unite("answer", q2:q8, sep ="", na.rm = TRUE) %>% 
  #filter out non-response for q2-q8
  filter(answer != "") %>% 
  # combine categories to meet assumptions of test
  mutate(answer = case_when( 
    answer == "Strongly opposed" ~"Opposed",
    answer == "Somewhat opposed" ~ "Opposed",
    answer == "Neither opposed nor in support" ~ "Neither opposed nor in support",
    answer == "Somewhat supportive" ~ "Supportive",
    answer == "Strongly supportive" ~ "Supportive")) %>%
  mutate(answer = fct_relevel(answer, c("Opposed", "Neither opposed nor in support", "Supportive"))) %>% 
  mutate(q15 = fct_explicit_na(q15, na_level = "0 days")) %>% 
  mutate(q15 = fct_relevel(q15, c("0 days", "1-5 days", "5-10 days", "10-15 days", "15+ days"))) %>% 
  group_by(q15, answer) %>% 
  summarize(weight_count = round(sum(weight),)) %>% 
  mutate(proportion = (weight_count/sum(weight_count)))

ocean_x2_table <- ocean_x2 %>% 
  pivot_wider(names_from = answer,
              values_from = weight_count) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  column_to_rownames(var = "q15")
```

```{r}
# Chi-square test for independence: Is there an association between support for aquaculture and word choice?
ocean_result <- chisq.test(ocean_x2_table)
ocean_result

ocean_results_tidy <- tidy(ocean_result)
```
